% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{AcousticData}
\alias{AcousticData}
\title{StoX data type AcousticData}
\description{
Biotic data read from biotic xml files.
}
\details{
This StoX data type is produced by \code{\link{ReadAcoustic}}, and contains one list per input acoustic file holding the tables read from each file, added a table named "metadata" holding the input file path and format. Currently supported are NMDEchosounder1 (\url{https://www.imr.no/formats/nmdechosounder/v1/}), and ICESAcoustic (\url{https://ices.dk/data/data-portals/Pages/acoustic.aspx}, click on "Acoustic data format" to download the format description).
}
\seealso{
\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxData}}
}
