% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBiotic.R
\name{AddStoxBioticVariables}
\alias{AddStoxBioticVariables}
\title{Merge StoxBioticData}
\usage{
AddStoxBioticVariables(
  StoxBioticData,
  BioticData,
  TableName,
  VariableName,
  cores = NULL
)
}
\arguments{
\item{StoxBioticData}{A list of StoX biotic data (StoX data type \code{\link{StoxBioticData}}).}
}
\value{
An object of StoX data type \code{\link{MergedStoxBioticData}}.
}
\description{
Merge StoxBioticData
}
