% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBiotic.R
\name{AddToStoxBiotic}
\alias{AddToStoxBiotic}
\title{Add variables to StoxBioticData from BioticData}
\usage{
AddToStoxBiotic(
  StoxBioticData,
  BioticData,
  VariableNames = character(),
  NumberOfCores = integer()
)
}
\arguments{
\item{StoxBioticData}{A list of StoX biotic data (StoX data type \code{\link{StoxBioticData}}).}

\item{BioticData}{A list of biotic data (StoX data type \code{\link{BioticData}}), one element for each input biotic file.}

\item{VariableNames}{A character vector with names of the variables to add from the \code{BioticData}.}

\item{NumberOfCores}{Overrides multi-core auto detection (default).}
}
\value{
An object of StoX data type \code{\link{StoxBioticData}}.
}
\description{
Add variables to StoxBioticData from BioticData
}
