% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxExport.R
\name{aggregateSaInICESAcoustic}
\alias{aggregateSaInICESAcoustic}
\title{Aggregate sa in ICESAcoustic}
\usage{
aggregateSaInICESAcoustic(ICESAcousticData)
}
\arguments{
\item{ICESAcousticData}{A \code{\link{ICESAcousticData}} object obtained from an ICES acoustic XML format file.}
}
\value{
List of string matrices in the ICES acoustic CSV format.
}
\description{
Aggregate sa in ICESAcoustic
}
