% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwardCompatibility.R
\name{backwardCompatibility_RstoxData}
\alias{backwardCompatibility_RstoxData}
\title{Backward compabitibility actions:}
\usage{
backwardCompatibility_RstoxData
}
\description{
Backward compabitibility actions:
}
