% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{BioticAcousticConversion}
\alias{BioticAcousticConversion}
\title{Conversion table for \code{\link{AcousticData}}.}
\description{
Conversion table for \code{\link{AcousticData}}.
}
\details{
This StoX data type is produced by \code{\link{DefineAcousticVariableConversion}}, and contains the columns VariableName, Value and NewValue.
}
\seealso{
\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxData}}
}
