% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineAndUpdateVariables.R
\name{ConvertAcousticVariables}
\alias{ConvertAcousticVariables}
\title{Convert Acoustic variables}
\usage{
ConvertAcousticVariables(
  AcousticData,
  ConversionMethod = c("Table", "PreDefined"),
  VariableConversionTable = data.table::data.table(),
  AcousticVariableConversion
)
}
\arguments{
\item{AcousticData}{Input \code{\link{AcousticData}} data.}

\item{ConversionMethod}{Character: A string naming the method to use, one of "Table", for providing the old and new values of the variables in the table \code{VariableConversionTable}; and "PreDefined" for providing the table using the argument \code{AcousticVariableConversion}.}

\item{VariableConversionTable}{A table of the columns \code{VariableName}, \code{Value} and \code{NewValue}, specifying the conversion from Value to NewValue for any number of variables of any number of tables of \code{\link{AcousticData}}.}

\item{AcousticVariableConversion}{The \code{\link{AcousticVariableConversion}} process data.}
}
\value{
A \code{\link{AcousticData}} object.
}
\description{
This function converts specific vaiables of specific tables of AcousticData to used defined new values.
}
