% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeXmlFile.R
\name{convertBioticFile}
\alias{convertBioticFile}
\title{Converts biotic}
\usage{
convertBioticFile(
  sourceFile,
  targetFile,
  targetFormat = "nmdbioticv3.0.xsd",
  overwrite = F
)
}
\arguments{
\item{sourceFile}{path to file that should be converted}

\item{targetFile}{path that the converted files should be written to}

\item{targetFormat}{name of xsdObject specifying the format of the target file. One of names(RstoxData::xsdObjects).}

\item{overwrite}{if TRUE any existing file in 'targetFile' will be overwritten.}
}
\description{
Converts between compatible nmdbiotic versions
}
\details{
Different versions of the nmdbiotic format are compatible and can be converted with this function
 if they have the same major release number. So that v3.1 is compatible with v3.0, but not with version 1.4
 
 Compatibility is not enforced by this function, and conversion will be attempted without checking the version of 'source'
}
