% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineAndUpdateVariables.R
\name{ConvertBioticVariables}
\alias{ConvertBioticVariables}
\title{Convert Biotic variables}
\usage{
ConvertBioticVariables(
  BioticData,
  ConversionMethod = c("Table", "PreDefined"),
  VariableConversionTable = data.table::data.table(),
  BioticVariableConversion
)
}
\arguments{
\item{BioticData}{Input \code{\link{BioticData}} data.}

\item{ConversionMethod}{Character: A string naming the method to use, one of "Table", for providing the old and new values of the variables in the table \code{VariableConversionTable}; and "PreDefined" for providing the table using the argument \code{BioticVariableConversion}.}

\item{VariableConversionTable}{A table of the columns \code{VariableName}, \code{Value} and \code{NewValue}, specifying the conversion from Value to NewValue for any number of variables of any number of tables of \code{\link{BioticData}}.}

\item{BioticVariableConversion}{The \code{\link{BioticVariableConversion}} process data.}
}
\value{
A \code{\link{BioticData}} object.
}
\description{
This function converts specific vaiables of specific tables of BioticData to used defined new values.
}
