% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineAndUpdateVariables.R
\name{ConvertStoxAcousticVariables}
\alias{ConvertStoxAcousticVariables}
\title{Convert StoxAcoustic variables}
\usage{
ConvertStoxAcousticVariables(
  StoxAcousticData,
  ConversionMethod = c("Table", "PreDefined"),
  VariableConversionTable = data.table::data.table(),
  StoxAcousticVariableConversion
)
}
\arguments{
\item{StoxAcousticData}{Input \code{\link{StoxAcousticData}} data.}

\item{ConversionMethod}{Character: A string naming the method to use, one of "Table", for providing the old and new values of the variables in the table \code{VariableConversionTable}; and "PreDefined" for providing the table using the argument \code{AcousticVariableConversion}.}

\item{VariableConversionTable}{A table of the columns \code{TableName}, \code{VariableName}, \code{Value} and \code{NewValue}, specifying the conversion from Value to NewValue for any number of variables of any number of tables of \code{\link{AcousticData}}.}

\item{StoxAcousticVariableConversion}{The (optional) \code{\link{StoxAcousticVariableConversion}} process data.}
}
\value{
A \code{\link{StoxAcousticData}} object.
}
\description{
This function converts specific vaiables of specific tables of StoxAcousticData to used defined new values.
}
