% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineAndUpdateVariables.R
\name{ConvertStoxBioticVariables}
\alias{ConvertStoxBioticVariables}
\title{Convert StoxBiotic variables}
\usage{
ConvertStoxBioticVariables(
  StoxBioticData,
  ConversionMethod = c("Table", "PreDefined"),
  VariableConversionTable = data.table::data.table(),
  StoxBioticVariableConversion
)
}
\arguments{
\item{StoxBioticData}{Input \code{\link{StoxBioticData}} data.}

\item{ConversionMethod}{Character: A string naming the method to use, one of "Table", for providing the old and new values of the variables in the table \code{VariableConversionTable}; and "PreDefined" for providing the table using the argument \code{BioticVariableConversion}.}

\item{VariableConversionTable}{A table of the columns \code{TableName}, \code{VariableName}, \code{Value} and \code{NewValue}, specifying the conversion from Value to NewValue for any number of variables of any number of tables of \code{\link{BioticData}}.}

\item{StoxBioticVariableConversion}{The (optional) \code{\link{StoxBioticVariableConversion}} process data.}
}
\value{
A \code{\link{StoxBioticData}} object.
}
\description{
This function converts specific vaiables of specific tables of StoxBioticData to used defined new values.
}
