% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdirParsing.R
\name{convertToLssData}
\alias{convertToLssData}
\title{Convert landings}
\usage{
convertToLssData(LandingData)
}
\arguments{
\item{LandingData}{\code{\link[RstoxData]{LandingData}} landings to convert}
}
\value{
converted landings. See \code{\link[RstoxData]{readLssFile}} for format
}
\description{
Convert landings to format format provided by \code{\link[RstoxData]{readLssFile}}.
}
\details{
When converting from lss:
 All columns of 'lssLandings' are populated, except 'salgsdato' and 'versjonstidspunkt'.
 All columns of \code{\link[RstoxData]{LandingData}} are converted, except 'DokumentFormulardato' and 'DokumentElektroniskDato'.
}
