% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{convertUnits}
\alias{convertUnits}
\title{convert units}
\usage{
convertUnits(value, unit, desired, conversionTable = RstoxData::StoxUnits)
}
\arguments{
\item{value}{to be expressed in different unit}

\item{unit}{symbol of unit for value}

\item{desired}{symbol of unit desired after conversion}

\item{conversionTable}{formatted as \code{\link[RstoxData]{StoxUnits}}}
}
\value{
value expressed in 'desired' unit.
}
\description{
Express values in a different unit.
}
\details{
The argument 'conversionTable' defines valid units and their symbols.
 This defaults to \code{\link[RstoxData]{StoxUnits}}.
}
\examples{
 convertUnits(c(1,2), "km", "nmi")
 convertUnits(10, "t", "g")
}
