% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Copy.R
\name{CopyAcoustic}
\alias{CopyAcoustic}
\title{Copy a variable of AcousticData}
\usage{
CopyAcoustic(
  AcousticData,
  FromVariable = character(),
  ToVariable = character(),
  Overwrite = FALSE,
  PreserveClass = TRUE
)
}
\arguments{
\item{AcousticData}{\code{\link{AcousticData}}.}

\item{FromVariable}{The name of the variable to copy.}

\item{ToVariable}{The name of the (possibly existing) variable insert the values in \code{FromVariable} to.}

\item{Overwrite}{Logical: If TRUE overwrite the existing variable, if \code{ToVariable} is the name of an existing variable in the same table as \code{FromVariable}.}

\item{PreserveClass}{Logical: Only used if \code{Overwrite} is TRUE. If \code{PreserveClass} is TRUE (the default) do not convert the class of the variable given by \code{ToVariable} to the class of the variable given by \code{FromVariable}.}
}
\value{
A \code{\link{AcousticData}} object.
}
\description{
This function copies a variable to another (possibly existing) variable of \code{\link{AcousticData}} .
}
