% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Copy.R
\name{CopyBiotic}
\alias{CopyBiotic}
\title{Copy a variable of BioticData}
\usage{
CopyBiotic(
  BioticData,
  FromVariable = character(),
  ToVariable = character(),
  Overwrite = FALSE,
  PreserveClass = TRUE
)
}
\arguments{
\item{BioticData}{\code{\link{BioticData}}.}

\item{FromVariable}{The name of the variable to copy.}

\item{ToVariable}{The name of the (possibly existing) variable insert the values in \code{FromVariable} to.}

\item{Overwrite}{Logical: If TRUE overwrite the existing variable, if \code{ToVariable} is the name of an existing variable in the same table as \code{FromVariable}.}

\item{PreserveClass}{Logical: If TRUE (the default) do not convert the class of the variable given by \code{ToVariable} to the class of the variable given by \code{FromVariable}.}
}
\value{
A \code{\link{BioticData}} object.
}
\description{
This function copies a variable to another (possibly existing) variable of \code{\link{BioticData}} .
}
