% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.R
\name{cropAcoustic}
\alias{cropAcoustic}
\title{Write a cropped copy of an acoustic file}
\usage{
cropAcoustic(
  filePath,
  logsToKeep = NULL,
  newFilePath = NULL,
  ow = FALSE,
  suffix = "_croppedLogs."
)
}
\arguments{
\item{filePath}{Character: The path to the acoustic file.}

\item{logsToKeep}{Numeric: The indiced of the logs to keep. NULL implies to keep all logs.}

\item{newFilePath}{Optional, character: The path to the new file, defaulted to \code{filePath} with the suffix before file extension}

\item{ow}{Logical: If TRUE overwrite the \code{newFilePath}.}

\item{suffix}{Character: The suffix to append to the \code{filePath} before the file extension, in the case that \code{newFilePath} is not specified}
}
\value{
The path to the cropped file.
}
\description{
This function writes a copy of an acoustic file in either the NMDEchosounder og ICESAcoustic format, keeping only the logs specified by the user.
}
\examples{
exampleFile <- system.file("testresources","ICES_Acoustic_2.xml", package="RstoxData")
acousticData <- ReadAcoustic(exampleFile)
newFilePath <- tempfile(fileext = ".xml")
cropAcoustic(exampleFile, logsToKeep = 2, newFilePath = newFilePath)
croppedAcousticData <- ReadAcoustic(newFilePath)

}
