% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{DataTypes}
\alias{DataTypes}
\title{StoX data types of the RstoxData package}
\description{
StoX data types are the data types used to transfer data and information between processes in a StoX estimation model. The data types are divided into two types, the \code{\link{ModelData}} and \code{\link{ProcessData}}.
}
