% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineAndUpdateVariables.R
\name{DefineBioticVariableConversion}
\alias{DefineBioticVariableConversion}
\title{Biotic variable conversion}
\usage{
DefineBioticVariableConversion(processData, FileName, UseProcessData = FALSE)
}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{FileName}{A file from which to read the \code{VariableConversionTable}.}

\item{UseProcessData}{Logical: If TRUE use the existing function output in the process.}
}
\value{
A \code{\link{BioticVariableConversion}} object.
}
\description{
This function reads a file holding a table for converting values of one or more variables to new values in \code{\link{BioticData}}. The actual conversion is done by \code{\link{ConvertBioticVariables}}.
}
