% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{do.call_robust}
\alias{do.call_robust}
\title{do.call which ignores non-relevant parameters}
\usage{
do.call_robust(
  what,
  args,
  quote = FALSE,
  envir = parent.frame(),
  keep.unnamed = FALSE
)
}
\arguments{
\item{what}{either a function or a non-empty character string naming the
    function to be called.}

\item{args}{a \emph{list} of arguments to the function call.  The
    \code{names} attribute of \code{args} gives the argument names.}

\item{quote}{a logical value indicating whether to quote the
    arguments.}

\item{envir}{an environment within which to evaluate the call.  This
    will be most useful if \code{what} is a character string and
    the arguments are symbols or quoted expressions.}

\item{keep.unnamed}{Logical: If TRUE keep all unnamed arguments.}
}
\description{
do.call which ignores non-relevant parameters
}
