% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-ecosystemsurvey_example.R
\docType{data}
\name{ecosystemsurvey_example}
\alias{ecosystemsurvey_example}
\title{Data from Barents Sea NOR-RUS ecosystem cruise in autumn}
\format{
\code{\link[RstoxData]{BioticData}}
}
\usage{
data(ecosystemsurvey_example)
}
\description{
Example data from the ecosystem survey.
Data is sourced from NMDbiotic.
}
\examples{
data(ecosystemsurvey_example)
}
\keyword{datasets}
