% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{FilterAcoustic}
\alias{FilterAcoustic}
\title{Filter (raw) Acoustic data}
\usage{
FilterAcoustic(AcousticData, FilterExpression, FilterUpwards = FALSE)
}
\arguments{
\item{AcousticData}{Input \code{\link{AcousticData}} data.}

\item{FilterExpression}{Filter expression in list of strings. The name of the list and structures should be identical to the names of the input data list.}

\item{FilterUpwards}{Whether the filter action will propagate in the upwards direction. Default to FALSE.}
}
\value{
An object of filtered data in the same format as the input data.
}
\description{
Filters \code{\link{AcousticData}}.
}
