% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{FilterBiotic}
\alias{FilterBiotic}
\title{Filter (raw) Biotic data}
\usage{
FilterBiotic(BioticData, FilterExpression = list(), FilterUpwards = FALSE)
}
\arguments{
\item{BioticData}{Input \code{\link{BioticData}} data.}

\item{FilterExpression}{Filter expression given as a list of strings. The name of the list and structures should be identical to the names of the input data list. To extract or exclude missing values (NAs) use the operator \code{\%in\%} or the special operator \code{\%notin\%}, which is defined in RstoxData. For a numeric filter, in the current StoX GUI, NA must be added manually in the text field of the "Expression table dialog" (the "Query builder dialog" will only show numeric values and will show an empty fields if the filter contains NA). Possible values are provided in the filter expression builder in the StoX GUI for character values, integer values, and numeric values that are all whole numbers, but not for keys (CruiseKey, StationKey, HaulKey, SpeciesCategoryKey, SampleKey and IndividualKey in \code{\link{FilterStoxBiotic}}, and CruiseKey, LogKey, BeamKey, AcousticCategoryKey, ChannelReferenceKey and NASCKey in \code{\link{FilterStoxAcoustic}}).}

\item{FilterUpwards}{Whether the filter action will propagate in the upwards direction. Default to FALSE. Use this option with caution, particularly for swept-area survey estimates, where setting \code{FilterUpwards} to TRUE could affect the estimated mean density.}
}
\value{
An object of filtered data in the same format as the input data.
}
\description{
Filters \code{\link{BioticData}}.
}
