% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{filterData}
\alias{filterData}
\title{Run filter on any StoX related data source}
\usage{
filterData(
  inputData,
  filterExpression,
  propagateDownwards = TRUE,
  propagateUpwards = FALSE,
  useXsd = FALSE
)
}
\arguments{
\item{inputData}{An input data. Can be a list of biotic data (StoX data type \code{\link{BioticData}}), list of acoustic data, StoxBiotic data, or StoxAcoustic data.}

\item{filterExpression}{Filter expression in list of strings. The name of the list and structures should be identical to the names of the input data list.}

\item{propagateDownwards}{Whether the filter action will propagate in the downwards direction. Default to TRUE.}

\item{propagateUpwards}{Whether the filter action will propagate in the upwards direction. Default to FALSE.}

\item{useXsd}{logical determining if treeStruct should be read from xsd provided in metadata table. See details.}
}
\value{
An object of filtered data in the same format as the input data.
}
\description{
Applies filters to hierarchical data formats
}
\details{
Applies filters to a set of table that are hierarchically related, and provides options for
propagating filters up or down in the data hierarchy.

Filters are specified as strings encoding logical expressions in standard R syntax, each associated to one of
the tables in the input.

After a filter is applied to a table, data that corresponds to the removed in other tables will be removed
as specified by the options:
\itemize{
\item propagateDownwards removes all data lower in the hierarchy that was linked to data explicitly removed by filter.
\item propagateUpwards removes all data higher in the hierarchy that is not linked with anything at the levels where explicitly filtered data has been removed.
}

The argument 'useXsd' specifies if the hierarchical relation between tables in 'inputData', should be inferred
from \code{\link[RstoxData]{xsdObjects}} via the conventional use of a metadata table on the 'inputData'

If useXsd is FALSE, each table in the hierarchy is assumed to be strictly lower in the hierarchy than any table preceding it,
and any table called 'metadata' is ignored.

Column names are assumed to be unique across all levels of the data hierarchy, except that each table is expected to
repeated the key-columns of all its ancestors.
}
\examples{
 filenames <- system.file("testresources",
       "biotic3.1_w_ageandprey.xml", package="RstoxData")
 inputData <- RstoxData:::ReadBiotic(filenames)
 
 #filters are nested list matching the structure of inputData
 filterExpression <- list()
 filterExpression$`biotic3.1_w_ageandprey.xml`$agedetermination <- c(
     'age < 10'
 )
 
 filteredData <- RstoxData::filterData(inputData, filterExpression, useXsd=TRUE)
 nrow(filteredData$biotic3.1_w_ageandprey.xml$fishstation)
 filteredPropUp <- RstoxData::filterData(inputData, filterExpression, propagateUpwards = TRUE, useXsd=TRUE)
 nrow(filteredPropUp$biotic3.1_w_ageandprey.xml$fishstation)

}
