% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{FilterICESAcoustic}
\alias{FilterICESAcoustic}
\title{Filter ICESAcoustic data}
\usage{
FilterICESAcoustic(
  ICESAcousticData,
  FilterExpression = list(),
  FilterUpwards = FALSE
)
}
\arguments{
\item{ICESAcousticData}{Input \code{\link{ICESAcousticData}} data.}

\item{FilterExpression}{Filter expression in list of strings. The name of the list and structures should be identical to the names of the input data list.}

\item{FilterUpwards}{Whether the filter action will propagate in the upwards direction. Default to FALSE.}
}
\value{
An object of filtered data in the same format as the input data.
}
\description{
Filters \code{\link{ICESAcousticData}}.
}
