% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{FilterStoxAcoustic}
\alias{FilterStoxAcoustic}
\title{Filter StoxAcoustic data}
\usage{
FilterStoxAcoustic(StoxAcousticData, FilterExpression, FilterUpwards = FALSE)
}
\arguments{
\item{StoxAcousticData}{Input \code{\link{StoxAcousticData}} data.}

\item{FilterExpression}{Filter expression in list of strings. The name of the list and structures should be identical to the names of the input data list.}

\item{FilterUpwards}{Whether the filter action will propagate in the upwards direction. Default to FALSE.}
}
\value{
An object of filtered data in the same format as the input data.
}
\description{
Filters \code{\link{StoxAcousticData}}.
}
