% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{FilterStoxLanding}
\alias{FilterStoxLanding}
\title{Filter StoxLanding data}
\usage{
FilterStoxLanding(StoxLandingData, FilterExpression)
}
\arguments{
\item{StoxLandingData}{Input \code{\link{StoxLandingData}} data.}

\item{FilterExpression}{Filter expression in list of strings. The name of the list and structures should be identical to the names of the input data list.}
}
\value{
An object of filtered data in the same format as the input data.
}
\description{
Filters \code{\link{StoxLandingData}}.
}
