% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{findUnit}
\alias{findUnit}
\title{Look up unit}
\usage{
findUnit(
  quantity,
  shortname = NULL,
  symbol = NULL,
  name = NULL,
  unitTable = RstoxData::StoxUnits
)
}
\arguments{
\item{quantity}{The quantity to look up units for}

\item{shortname}{shortname of the desired unit}

\item{symbol}{Symbol of the desired unit}

\item{name}{name of the desired unit}

\item{unitTable}{formatted as \code{\link[RstoxData]{StoxUnits}}}
}
\value{
valid id in unitTable
}
\description{
Find the unit identifier given quantity and either symbol, shortname or name
}
\details{
This is useful for looking up correct unit identifiers from user input.
 Provide either symbol, shortname or name, do not provide more than one.
 
 The argument 'conversionTable' defines valid units and their symbols.
 This defaults to \code{\link[RstoxData]{StoxUnits}}.
}
\examples{
 findUnit("mass", getUnitOptions("mass")[1])
 findUnit("length", "m")
}
\seealso{
\code{\link[RstoxData]{getUnitOptions}} for listing all valid options for a quantity.
}
