% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{firstClass}
\alias{firstClass}
\title{Get the first element returned by class(), and translate "double" to "numeric".}
\usage{
firstClass(x)
}
\arguments{
\item{x}{An R object.}
}
\description{
Get the first element returned by class(), and translate "double" to "numeric".
}
