% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{general_arguments}
\alias{general_arguments}
\title{General parameters of RstoxFramework.}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{UseProcessData}{Logical: If TRUE use the existing function output in the process.}
}
\description{
All functions referring to a project, a model, a process or an output table use the same parameters, listed here.
}
