% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{getKeys}
\alias{getKeys}
\title{Get keys from the data}
\usage{
getKeys(data)
}
\arguments{
\item{data}{An StoX data object}
}
\description{
Look for the metadata list element specifying the XSD, or estimate the keys based on intersection between tables, in which case the keys of the last table is set to the keys of the parent plus the next field. For more control of keys, please use getRstoxDataDefinitions("keys") and feed this directly into translateVariables().
}
