% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{getNumberOfCores}
\alias{getNumberOfCores}
\title{Pick a suitable number of cores}
\usage{
getNumberOfCores(NumberOfCores = NULL, n = NULL)
}
\arguments{
\item{NumberOfCores}{The number of cores to use (defaulted to 1), truncated to the number of avaliable cores.}

\item{n}{Optional length of the data to apply parallel processing to.}
}
\value{
The number of cores to apply.
}
\description{
Pick a suitable number of cores
}
