% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Definitions.R
\name{getRstoxDataDefinitions}
\alias{getRstoxDataDefinitions}
\title{Get RstoxData definitions}
\usage{
getRstoxDataDefinitions(name = NULL, ...)
}
\arguments{
\item{name}{An optional string vector denoting which definitions to extract.}

\item{...}{values overriding the values of definitions.}
}
\value{
A list of definitions.
}
\description{
This function gets vital definitions from the RstoxData environment.
}
\examples{
getRstoxDataDefinitions()

}
