% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{getStoxKeys}
\alias{getStoxKeys}
\title{Get the keys of a StoX format}
\usage{
getStoxKeys(
  StoxDataType = c("StoxBiotic", "StoxAcoustic"),
  level = NULL,
  keys.out = c("all", "only.present", "all.but.present")
)
}
\arguments{
\item{StoxDataType}{The name of the StoX format (only StoxBiotic implemented yet).}

\item{level}{The name of the level/table to get keys for.}

\item{keys.out}{Specification of what to return. One of "all", to return all keys of the level; "only.present", to return only the key of the \code{level}; and "all.but.present", to return all keys except the present key.}
}
\description{
Get the keys of a StoX format
}
