% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{getUnitOptions}
\alias{getUnitOptions}
\title{Get available units}
\usage{
getUnitOptions(quantity, unitTable = RstoxData::StoxUnits)
}
\arguments{
\item{quantity}{quantity to get units for, such as 'mass', 'length' etc.}

\item{unitTable}{formatted as \code{\link[RstoxData]{StoxUnits}}}
}
\value{
a character vector with availble unit symbols.
}
\description{
Get the unit symbmols that are available for a given quantity.
 The available quntities and units are defined by the argument unitTable
 which defaults to RstoxData::StoxUnits
}
\examples{
 print(getUnitOptions("mass"))
}
