% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{hasUnit}
\alias{hasUnit}
\title{Does the input have unit?}
\usage{
hasUnit(
  value,
  property = c("id", "shortname", "symbol", "name"),
  unitTable = RstoxData::StoxUnits
)
}
\arguments{
\item{value}{value to read unit from}

\item{property}{the property of the unit that is to be returned (e.g. 'symbol' or 'name')}

\item{unitTable}{formatted as \code{\link[RstoxData]{StoxUnits}}}
}
\value{
TRUE if the \code{value} has unit, FALSE if not.
}
\description{
Get unit id, shortname, symbol or name of a value in accordance with StoX convention.
}
\examples{
 dt <- data.table::data.table(weight=c(1000,1200))
 dt$weight <- setUnit(dt$weight, "mass-g")
 print(hasUnit(dt$weight))
 print(hasUnit(1))
}
