% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxExport.R
\name{ICESAcoustic}
\alias{ICESAcoustic}
\title{Convert AcousticData to ICESAcousticData.}
\usage{
ICESAcoustic(AcousticData)
}
\arguments{
\item{AcousticData}{A \code{\link{AcousticData}} object from an ICES acoustic XML format file.}
}
\value{
An \code{\link{ICESAcousticData}} object.
}
\description{
Note that this function only supports \code{\link{AcousticData}} object created from reading an ICES acoustic XML files.
}
