% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ICESAcousticData}
\alias{ICESAcousticData}
\title{StoX data type ICESAcousticData}
\description{
Acoustic data stored in the ICESAcoustic (CSV) format.
}
\details{
This StoX data type is produced by \code{\link{ICESAcoustic}}, and contains one list per input biotic file read to produec the input to \code{\link{ICESAcoustic}}, each holding the tables Instrument, Calibration, DataAcquisition, DataProcessing, Cruise and Data (here Data is a table merged from Log, Sample and Data of the ICESAocustic xml format). Each file read to produec the input to \code{\link{ICESAcoustic}}
}
\seealso{
\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxData}}
}
