% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ICESBioticData}
\alias{ICESBioticData}
\title{StoX data type ICESBioticData}
\description{
Biotic data stored in the ICESBiotic (CSV) format.
}
\details{
This StoX data type is produced by \code{\link{ICESBiotic}}, and contains one list per input biotic file read to produec the input to \code{\link{ICESBiotic}}, each holding the tables Cruise, Haul, Catch and Biology, in that hierarchical order.
}
\seealso{
\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxData}}
}
