% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxExport.R
\name{ICESDatras}
\alias{ICESDatras}
\title{Convert BioticData to ICESDatras format}
\usage{
ICESDatras(BioticData)
}
\arguments{
\item{BioticData}{a \code{BioticData} object from an XML file with NMD biotic version 3 format.}
}
\value{
An \code{\link{ICESDatrasData}} object.
}
\description{
Given an \code{\link{BioticData}} object, this function converts to ICESDatras format. Note that this function only supports
\code{\link{BioticData}} NMDBiotic version > 3 XML files.
}
