% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxExport.R
\name{ICESDatsusc}
\alias{ICESDatsusc}
\title{Convert BioticData to ICESDatsuSC format}
\usage{
ICESDatsusc(BioticData)
}
\arguments{
\item{BioticData}{a \code{BioticData} object from an XML file with NMD biotic version 3 format.}
}
\value{
An \code{\link{ICESDatsuscData}} object.
}
\description{
Given an \code{\link{BioticData}} object, this function converts to ICESDatsuSC format. Note that this function only supports
\code{\link{BioticData}} NMDBiotic version > 3 XML files.
}
