% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ICESDatsuscData}
\alias{ICESDatsuscData}
\title{StoX data type ICESDatsuscData}
\description{
Biotic data stored in the ICESDatsusc (CSV) format.
}
\details{
This StoX data type is produced by \code{\link{ICESDatsusc}}, and contains one list per input biotic file read to produec the input to \code{\link{ICESDatusc}}, each holding the tables HH, HL and CA, in that hierarchical order.
}
\seealso{
\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxData}}
}
