% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoxLanding.R
\name{is.LandingData}
\alias{is.LandingData}
\title{Check if argument is LandingData}
\usage{
is.LandingData(LandingData)
}
\arguments{
\item{LandingData}{argument to be checked for data conformity}
}
\value{
logical, TRUE if argument conformed to specification for \code{\link[RstoxData]{LandingData}}
}
\description{
Checks if argument conforms to specification for \code{\link[RstoxData]{LandingData}}
}
