% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoxLanding.R
\name{is.StoxLandingData}
\alias{is.StoxLandingData}
\title{Check if argument is StoxLandingData}
\usage{
is.StoxLandingData(StoxLandingData)
}
\arguments{
\item{StoxLandingData}{argument to be checked for data conformity}
}
\value{
logical, TRUE if argument conformed to specification for \code{\link[RstoxData]{StoxLandingData}}
}
\description{
Checks if argument conforms to specification for \code{\link[RstoxData]{StoxLandingData}}
}
