% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{LandingData}
\alias{LandingData}
\title{LandingData}
\description{
LandingData
}
\section{Data}{

One entry 'Seddellinje' is one line of a sales-note or landing-note. 
These are issued as fish is landed, and a complete set of these for a period
can be considered a census of all first hand sale of fish sold from Norwegian vessels.
}

\section{Format}{

list() of \code{\link[data.table]{data.table}} 
representing the different complexTypes in namespace http://www.imr.no/formats/landinger/v2
For ease of merging: all top level attributes are repeated for all tables. And all line-identifying variables are included as top-level attributes.
}

\seealso{
\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxData}}
}
