% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{lapplyOnCores}
\alias{lapplyOnCores}
\title{Run a function on all elements of x on one or more cores}
\usage{
lapplyOnCores(x, FUN, NumberOfCores = integer(), ...)
}
\arguments{
\item{x}{An object to apply \code{FUN} to.}

\item{FUN}{The function to apply.}

\item{NumberOfCores}{The number of cores to use, defaulted to detect the available number of cores, but never to run on more cores than the number of elements of \code{x}.}

\item{...}{Additional arguments to \code{FUN}.}
}
\value{
A list of outputs from \code{FUN}.
}
\description{
Run a function on all elements of x on one or more cores
}
