% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{match_arg_informative}
\alias{match_arg_informative}
\title{Informative match.arg}
\usage{
match_arg_informative(
  arg,
  choices,
  several.ok = FALSE,
  arg_name = substitute(arg),
  ignore.case = FALSE
)
}
\arguments{
\item{arg}{a character vector (of length one unless \code{several.ok}
    is \code{TRUE}) or \code{NULL} which means to take \code{choices[1]}.}

\item{choices}{a character vector of candidate values, often missing, see
    \sQuote{Details}.}

\item{several.ok}{logical specifying if \code{arg} should be allowed
    to have more than one element.}

\item{arg_name}{The name of the argument, interpreted by default.}

\item{ignore.case}{Logical: Should case be igngored.}
}
\description{
Informative match.arg
}
