% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeXmlFile.R
\name{MergeAcoustic}
\alias{MergeAcoustic}
\title{Merge AcousticData}
\usage{
MergeAcoustic(AcousticData)
}
\arguments{
\item{AcousticData}{\code{\link{AcousticData}}.}
}
\description{
Function to merge the tables of \code{\link{AcousticData}}.
}
