% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{mergeByIntersect}
\alias{mergeByIntersect}
\title{Merge two data tables by the intersect of the names}
\usage{
mergeByIntersect(x, y, ..., msg = FALSE)
}
\arguments{
\item{x, y}{Data tables of class \code{\link[data.table]{data.table}}.}

\item{...}{Various overrides.}

\item{msg}{Verbose message switch, default to \code{FALSE}.}
}
\value{
A merged data table.
}
\description{
Merge two data tables by the intersect of the names
}
