% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{mergeByKeys}
\alias{mergeByKeys}
\title{Merge two data tables by keys (variables ending with "Key")}
\usage{
mergeByKeys(x, y, toMergeFromY = NULL, replace = FALSE, ...)
}
\arguments{
\item{x, y}{Data tables of class \code{\link[data.table]{data.table}}).}
}
\value{
A merged data table.
}
\description{
Merge two data tables by keys (variables ending with "Key")
}
