% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{mergeByStoxKeys}
\alias{mergeByStoxKeys}
\title{Merge two data tables by StoX keys}
\usage{
mergeByStoxKeys(x, y, StoxDataType, toMergeFromY = NULL, replace = FALSE, ...)
}
\arguments{
\item{x, y}{Data tables of class \code{\link[data.table]{data.table}}).}
}
\value{
A merged data table.
}
\description{
Merge two data tables by StoX keys
}
