% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{mergeDataTables}
\alias{mergeDataTables}
\title{Merge list of data tables recursively}
\usage{
mergeDataTables(data, tableNames = NULL, output.only.last = FALSE, ...)
}
\arguments{
\item{data}{A list of data tables.}

\item{tableNames}{A character vector holding the names of the tables to merge.}

\item{output.only.last}{Only returns last merged table.}

\item{...}{Extra parameters that will be passed into \code{\link[data.table]{merge}}.}
}
\value{
A merged data table.
}
\description{
Merge list of data tables recursively
}
