% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{MergedStoxBioticData}
\alias{MergedStoxBioticData}
\title{StoX data type MergedStoxBioticData}
\description{
Merged \code{\link{StoxBioticData}}.
}
\details{
This StoX data type is produced by \code{\link{MergeStoxBiotic}}, and contains one merged table of \code{\link{StoxBioticData}}.
}
\seealso{
\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxData}}
}
