% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAcoustic.R
\name{MergeStoxAcoustic}
\alias{MergeStoxAcoustic}
\title{Merge StoxAcousticData}
\usage{
MergeStoxAcoustic(StoxAcousticData, TargetTable = "NASC")
}
\arguments{
\item{StoxAcousticData}{A list of StoX acoustic data (StoX data type \code{\link{StoxAcousticData}}).}

\item{TargetTable}{The name of the table up until which to merge (the default "NASC" implies merging all tables)}
}
\value{
An object of StoX data type \code{\link{MergeStoxAcousticData}}.
}
\description{
Merge StoxAcousticData
}
