% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{MergeStoxAcousticData}
\alias{MergeStoxAcousticData}
\title{StoX data type MergeStoxAcousticData}
\description{
Merged \code{\link{StoxAcousticData}}.
}
\details{
This StoX data type is produced by \code{\link{MergeStoxAcoustic}}, and contains one merged table of \code{\link{StoxAcousticData}}.
}
\seealso{
\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxData}}
}
