% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBiotic.R
\name{MergeStoxBiotic}
\alias{MergeStoxBiotic}
\title{Merge StoxBioticData}
\usage{
MergeStoxBiotic(StoxBioticData, TargetTable = "Individual")
}
\arguments{
\item{StoxBioticData}{A list of StoX biotic data (StoX data type \code{\link{StoxBioticData}}).}

\item{TargetTable}{The name of the table up until which to merge (the default "Individual" implies merging all tables)}
}
\value{
An object of StoX data type \code{\link{MergeStoxBioticData}}.
}
\description{
Merge StoxBioticData
}
