% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ModelData}
\alias{ModelData}
\title{Model data used by RstoxData}
\arguments{
\item{BioticData}{\code{\link{BioticData}}.}

\item{AcousticData}{\code{\link{AcousticData}}.}

\item{LandingData}{\code{\link{LandingData}}.}

\item{StoxBioticData}{\code{\link{StoxBioticData}}.}

\item{StoxAcousticData}{\code{\link{StoxAcousticData}}.}

\item{StoxLandingData}{\code{\link{StoxLandingData}}.}

\item{MergeStoxBioticData}{\code{\link{MergeStoxBioticData}}.}

\item{MergeStoxAcousticData}{\code{\link{MergeStoxAcousticData}}.}

\item{ICESBioticData}{\code{\link{ICESBioticData}}.}

\item{ICESAcousticData}{\code{\link{ICESAcousticData}}.}

\item{ICESDatrasData}{\code{\link{ICESDatrasData}}.}

\item{WriteICESBioticData}{\code{\link{WriteICESBioticData}}.}

\item{WriteICESAcousticData}{\code{\link{WriteICESAcousticData}}.}

\item{WriteICESDatrasData}{\code{\link{WriteICESDatrasData}}.}
}
\description{
The model data of a StoX model are the data generated during the model run based on input data and user settings and resources given in the project description (project.json file). Model data are transient and only exists from a process has been run until the project is closed.
}
