% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.R
\name{NMDBioticToICESBiotic}
\alias{NMDBioticToICESBiotic}
\title{Convert BioticData to ICESBiotic form.}
\usage{
NMDBioticToICESBiotic(
  BioticData,
  SurveyName = "NONE",
  Country = character(),
  Organisation = integer(),
  AllowRemoveSpecies = TRUE
)
}
\arguments{
\item{BioticData}{\code{\link{BioticData}}.}

\item{SurveyName}{A string naming the survey. Must be one of the names listed on \url{https://vocab.ices.dk/?ref=1453} or NONE (the default).}

\item{Country}{The country code as listed on \url{http://vocab.ices.dk/?ref=337}.}

\item{Organisation}{An integer code representing the organization running the cruise. See \url{https://vocab.ices.dk/?ref=1398} for a list of possible codes (e.g., Institute of Marine Research: 612).}

\item{AllowRemoveSpecies}{Logical: ICES submission will not allow the resulting CSV file to be uploaded if the file contains species not listed on \url{https://acoustic.ices.dk/Services/Schema/XML/SpecWoRMS.xml}. Setting this parameter to TRUE will remove the unlisted species records.}
}
\value{
An object of StoX data type \code{\link{BioticData}}.
}
\description{
This function converts BioticData read from files of format NMDBiotic v3 and v3.1 to ICESBiotic tables.
}
\details{
The ICESBiotic (name used by StoX) data format is defined on \url{https://ices.dk/data/data-portals/Pages/acoustic.aspx} (\href{https://ices.dk/data/Documents/Acoustic/ICES_Acoustic_data_format_description.zip}{Direct download link}).
}
\seealso{
\code{\link{ReadBiotic}}.
}
