% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwardCompatibility.R
\name{oldToNewTranslationList}
\alias{oldToNewTranslationList}
\title{Function to convert from the VarialbeName-Value-NewValue translation form to the form where variable names are column names.}
\usage{
oldToNewTranslationList(translationList)
}
\arguments{
\item{translationList}{A list of lists of each row of the Translation table, as read in using RstoxFramework:::readProjectDescription().}
}
\description{
Function to convert from the VarialbeName-Value-NewValue translation form to the form where variable names are column names.
}
