% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{orderRowsByKeys}
\alias{orderRowsByKeys}
\title{Order a StoX format by keys}
\usage{
orderRowsByKeys(data)
}
\arguments{
\item{data}{An object of type \code{\link{StoxAcousticData}} or \code{\link{StoxBioticData}}}
}
\description{
Order a StoX format by keys
}
