% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxExport.R
\name{PrepareWriteICESDatras}
\alias{PrepareWriteICESDatras}
\title{Prepare for WriteICESDatras}
\usage{
PrepareWriteICESDatras(
  ICESDatrasData,
  RoundDownLngtClass = TRUE,
  RoundingTable = data.table::data.table(),
  AggregateHLNoAtLngt = TRUE
)
}
\arguments{
\item{ICESDatrasData}{A \code{\link{ICESDatrasData}} object as returned from \code{\link{ICESDatras}}.}

\item{RoundDownLngtClass}{Logical: If TRUE round down LngtClass to the appropriate resolution given by the column LngtCode and the \code{RoundingTable}.}

\item{RoundingTable}{A table of two columns; LngtCode (character) and LengthResolution (numeric) giving the resolution to apply when rounding down LngtClass. This table is merged to the ICESDatrasData using the LngtCode as key.}

\item{AggregateHLNoAtLngt}{Logical: If TRUE aggrerate HLNoAtLngt for each combination of StNo, SpecCode, Sex, CatIdentifier and LngtClass.}
}
\value{
An \code{\link{ICESDatrasData}} object.
}
\description{
Depending on the model, there may be actions that shuold be taken on the ICESDatrasData before writing the data to file. Currently this includes rounding down LngtClass, and aggregating for any duplicated keys (typically due to translation of LngtCode to reduce resolution, as each combination of haul and species need the same resolution).
}
