% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{printErrorIDs}
\alias{printErrorIDs}
\title{Paste to a linespace + tab separeted string}
\usage{
printErrorIDs(errorIDs, collapse = "\\n\\t")
}
\arguments{
\item{errorIDs}{A vector of strings.}

\item{collapse}{The separator.}
}
\description{
Paste to a linespace + tab separeted string
}
