% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ProcessData}
\alias{ProcessData}
\title{Process data used in estimation models in StoX}
\description{
The process data of the RstoxData package.
}
\details{
\itemize{
\item{\code{\link{StoxBioticTranslation}}}
}
}
\seealso{
\code{\link{ModelData}} for model data types and \code{\link{DataTypes}} for all data types produced by \code{\link{RstoxData}}.
}
