% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoxFunctionAttributes.R
\docType{data}
\name{processPropertyFormats}
\alias{processPropertyFormats}
\title{Define the process property formats:}
\format{
An object of class \code{list} of length 7.
}
\usage{
processPropertyFormats
}
\description{
Define the process property formats:
}
\keyword{datasets}
