% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.R
\name{ReadAcoustic}
\alias{ReadAcoustic}
\title{Read acoustic XML files}
\usage{
ReadAcoustic(FileNames = character())
}
\arguments{
\item{FileNames}{The paths of the acoustic files.}
}
\value{
An object of StoX data type AcousticData: A list of a list of data.tables of the different levels of the input acoustic files.
}
\description{
This function reads multiple acoustic file to a list with a list of tables for each file.
}
\examples{
exampleFile <- system.file(
    "testresources","libas_ListUserFile20__L40.0-2259.9_small.xml", package="RstoxData")
acousticData <- ReadAcoustic(exampleFile)

}
\seealso{
\code{\link{ReadBiotic}} and \code{\link{readXmlFile}}.
}
