% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdirParsing.R
\name{readErsFile}
\alias{readErsFile}
\title{Parses logbooks (ERS)}
\usage{
readErsFile(file, encoding = "latin1")
}
\arguments{
\item{file}{path to file}

\item{encoding}{encoding for 'file'}
}
\value{
data.table() with logbooks
}
\description{
Parses electronic logbooks (ERS) from tabular format delivered by Directorate of Fisheries (FDIR)
}
\details{
The format is a pipe-separated format encoding aggregated ERS records (logbooks).
 It is provided to IMR on a regular basis from FDIR.
 Column headers are in Norwegian.
}
