% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineAndUpdateVariables.R
\name{RedefineStoxBiotic}
\alias{RedefineStoxBiotic}
\title{Redefine StoxBioticData variables by data from BioticData}
\usage{
RedefineStoxBiotic(
  StoxBioticData,
  BioticData,
  Redefinition = data.table::data.table()
)
}
\arguments{
\item{StoxBioticData}{An input of \link{ModelData} object}

\item{BioticData}{An input of \link{ModelData} object}

\item{Redefinition}{A table of the columns "VariableName", representing the variable to redefine; and "RedefineBy", representing the variable from BioticData to replace by.}
}
\value{
A \code{\link{StoxBioticData}} object.
}
\description{
This function redefines one or more columns of \code{\link{StoxBioticData}} by columns of \code{\link{BioticData}}.
}
