% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxExport.R
\name{RegroupLengthICESDatras}
\alias{RegroupLengthICESDatras}
\title{Regroup LngtCode ICESDatrasData}
\usage{
RegroupLengthICESDatras(
  ICESDatrasData,
  GroupingVariables = character(),
  AggregateHLNoAtLngt = TRUE,
  AggregationVariables = character()
)
}
\arguments{
\item{ICESDatrasData}{A \code{\link{ICESDatrasData}} object as returned from \code{\link{ICESDatras}}.}

\item{GroupingVariables}{A vector of variable names giving the variables for which to reduce resolution to the lowest resolution.}

\item{AggregateHLNoAtLngt}{Logical: If TRUE aggregate the variable HLNoAtLngt after regrouping lengths.}

\item{AggregationVariables}{A vector of variables of the HL table for which to aggregate individuals. Usually this should include exactly the following variables: "StNo", "SpecCode", "Sex", "CatIdentifier", "LngtClass".}
}
\value{
An \code{\link{ICESDatrasData}} object.
}
\description{
ICES Datras requires equal LngtCode per haul and species.
}
