% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxExport.R
\name{ReportICESDatras}
\alias{ReportICESDatras}
\title{Reports \code{ICESAcousticData} to a csv file for each input acoustic file used to create the \code{\link{ICESAcousticData}}}
\usage{
ReportICESDatras(ICESDatrasData)
}
\arguments{
\item{ICESAcousticData}{A \code{ICESAcousticData} object obtained from an ICES acoustic XML format file.}
}
\value{
List of string matrices in the ICES acoustic CSV format.
}
\description{
Reports \code{ICESAcousticData} to a csv file for each input acoustic file used to create the \code{\link{ICESAcousticData}}
}
