% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{roundSignifSPDF}
\alias{roundSignifSPDF}
\title{Function setting the precision of one data table:}
\usage{
roundSignifSPDF(SPDF, digits, signifDigits)
}
\arguments{
\item{SPDF}{A SpatialPolygonsDataFrame object.}

\item{digits}{The number of digits to round off to. See also \code{signifDigits}.}

\item{signifDigits}{The number of significant digits, used when numbers are low to guarantee precision.}
}
\description{
Function setting the precision of one data table:
}
