% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RstoxData-package.R
\docType{package}
\name{RstoxData}
\alias{RstoxData}
\alias{RstoxData-package}
\title{Tools to Read and Manipulate Fisheries Data}
\description{
Set of tools to read and manipulate various data formats for fisheries. Mainly catered towards scientific trawl survey sampling ('biotic') data, acoustic trawl data, and commercial fishing catch ('landings') data. Among the supported data formats are the data products from the Norwegian Institute Marine Research ('IMR') and the International Council for the Exploration of the Sea (ICES).
}
\details{
The RstoxData package contains functions for reading, filtering and writing biotic, acoustic and landing data as XML files. Filtering can be done by R syntax such as longitude > 10, or by pre defined functions such as inside(). On computers that return errors when trying to run the Rtools through RStudio (most institutional Windows machines), install the binary directly from https://github.com/StoXProject/RstoxData/releases. Download the newest RstoxData zip file, click the "Packages" tab -> "Install" -> "Install from:" "Package Archive File" -> "Install". If the installer does not complain, the package is installed correctly.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StoXProject/RstoxData}
  \item Report bugs at \url{https://github.com/StoXProject/RstoxData/issues}
}

}
\author{
\strong{Maintainer}: Ibrahim Umar \email{ibrahim.umar@hi.no}

Authors:
\itemize{
  \item Sindre Vatnehol
  \item Arne Johannes Holmin
  \item Edvin Fuglebakk
  \item Espen Johnsen
}

Other contributors:
\itemize{
  \item Norwegian Institute of Marine Research [copyright holder, funder]
}

}
