% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{sanitizeExpression}
\alias{sanitizeExpression}
\title{Sanitize an expression given in a string}
\usage{
sanitizeExpression(x)
}
\arguments{
\item{x}{A string}
}
\description{
If the input contains "system" or "system2", an error is issued.
}
