% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{setRstoxPrecisionLevel}
\alias{setRstoxPrecisionLevel}
\title{Round off to number of digits}
\usage{
setRstoxPrecisionLevel(x)
}
\arguments{
\item{data}{A list of data tables.}

\item{tableNames}{A character vector holding the names of the tables to merge.}

\item{output.only.last}{Only returns last merged table.}
}
\value{
A merged data table.
}
\description{
Round off to number of digits
}
