% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{setUnit}
\alias{setUnit}
\title{Set units of value}
\usage{
setUnit(value, id, conversionTable = RstoxData::StoxUnits, assertNew = FALSE)
}
\arguments{
\item{value}{value to set unit for.}

\item{id}{id of the desired unit for the column}

\item{conversionTable}{formatted as \code{\link[RstoxData]{StoxUnits}}}

\item{assertNew}{if True an error is raised if the unit is already set for 'value'}
}
\value{
converted value with the attribute 'Unit' set / altered.
}
\description{
Set unit of a value in accordance with StoX convention (adds the attribute 'stoxUnit').
 Converts unit if already set.
}
\details{
The argument 'conversionTable' defines valid units and their symbols.
 This defaults to \code{\link[RstoxData]{StoxUnits}}.

 Providing id as NA removes unit from value
}
\examples{
 dt <- data.table::data.table(weight=c(1000,1200))
 dt$weight <- setUnit(dt$weight, "mass-g")
 print(dt$weight)
 dt$weight <- setUnit(dt$weight, "mass-kg")
 print(dt$weight)
}
