% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAcoustic.r
\name{StoxAcoustic}
\alias{StoxAcoustic}
\title{Convert AcousticData to StoxAcousticData}
\usage{
StoxAcoustic(AcousticData = NULL)
}
\arguments{
<<<<<<< HEAD
\item{AcousticData}{A list of acoustic data (StoX data type \code{\link{AcousticData}}), one element for each input acoustic file.}
=======
\item{AcousticData}{The output from xmlreader.}
>>>>>>> ef61effe44e8eed48af94c00c38af3e3216de6c6
}
\value{
An object of StoX data type \code{\link{StoxAcousticData}}.
}
\description{
Convert AcousticData to StoxAcousticData
}
