% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAcoustic.R
\name{StoxAcousticStartMiddleStopDateTime}
\alias{StoxAcousticStartMiddleStopDateTime}
\title{Generate Start, Middle and Stop DateTime variables}
\usage{
StoxAcousticStartMiddleStopDateTime(StoxAcousticData)
}
\arguments{
\item{StoxAcousticData}{A list of StoX acoustic data (StoX data type \code{\link{StoxAcousticData}}).}
}
\value{
An object of StoX data type \code{\link{MergeStoxAcousticData}}.
}
\description{
Generate Start, Middle and Stop DateTime variables
}
