% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{StoxBioticData}
\alias{StoxBioticData}
\title{StoX data type StoxBioticData}
\description{
Biotic data stored in the StoxBiotic format, which contains the variables needed for most estimation models used by StoX.
}
\details{
This StoX data type is produced by \code{\link{StoxBiotic}}, and contains the tables Cruise, Station, Haul, SpeciesCategory, Sample and Individual in that hierarchical order.
}
\seealso{
\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxData}}
}
