% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{StoxBioticMapping}
\alias{StoxBioticMapping}
\title{Mapping from BioticData to StoxBiotic data for different input file formats}
\description{
The \code{\link{StoxBioticFormat}} contains a fixed mapping between variables from the different input biotic data formats to the variables of the  \code{\link{StoxBioticData}}. For variables that are not supported by \code{\link{StoxBiotic}} the function \code{\link{AddToStoxBiotic}} can be used to add these to the \code{StoxBioticData}. The function \code{\link{RedefineStoxBiotic}} can also be used to redefine variables in a \code{StoxBioticData} object. Both of these functions add data to the \code{StoxBioticData} object using the described in the Details.
}
\details{
The mapping from the different biotic data format to StoxBiotic is different for different levels Some levels are directly mapped to a levels in \code{StoxBioticData}, whereas other levels are split into two levels of the \code{StoxBioticData}.

The following table shows levels of different formats that are mapped directly, i.e., all variables of the source levels are mapped to the desination levels:
 
  \tabular{llllll}{
   \bold{Format} \tab \bold{Level} \tab \bold{DesinationLevel} \cr
icesBiotic \tab Cruise \tab Cruise \cr
icesBiotic \tab Biology \tab Individual \cr
nmdbioticv1.1 \tab mission \tab Cruise \cr
nmdbioticv1.1 \tab individual \tab Individual \cr
nmdbioticv1.4 \tab mission \tab Cruise \cr
nmdbioticv1.4 \tab individual \tab Individual \cr
nmdbioticv3 \tab mission \tab Cruise \cr
nmdbioticv3 \tab individual \tab Individual \cr
nmdbioticv3.1 \tab mission \tab Cruise \cr
nmdbioticv3.1 \tab individual \tab Individual \cr
}

\bold{Note that the age from the agedetermination level of nmdbioticv1.1, nmdbioticv1.4, nmdbioticv3 and nmdbioticv3.1 is selected based on the preferredagereading in the individual level and placed in the Individual level of \code{StoxBioticData}.}

The mapping from one level to two levels is shown in the table below. The parameter \code{SplitTableAllocation} in \code{\link{AddToStoxBiotic}} and \code{\link{RedefineStoxBiotic}} is used to determine the mapping. The columns of the mapping are as follows:

 \describe{
   \item{Format}{The biotic data format}
   \item{Level}{The table in the \code{BioticData}}
   \item{Variable}{The variable in the \code{BioticData}}
   \item{DesinationLevel_Default}{The table of the \code{StoxBioticData} to which the variable is placed by default (SplitTableAllocation = "Default")}
   \item{DesinationLevel_Lowest}{The table of the \code{StoxBioticData} to which the variable is placed when SplitTableAllocation = "Lowest"}
   \item{DesinationLevel_Highest}{The table of the \code{StoxBioticData} to which the variable is placed when SplitTableAllocation = "Highest"}
 }

\bold{Mapping from one level in \code{BioticData} to two levels in \code{StoxBioticData} }
\tabular{llllll}{
\bold{Format} \tab \bold{Level} \tab \bold{Variable} \tab \bold{DesinationLevel_Default} \tab \bold{DesinationLevel_Lowest} \tab \bold{DesinationLevel_Highest} \cr
icesBiotic \tab Haul \tab CruiseKey \tab Station \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab StationKey \tab Station \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab Gear \tab Haul \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab Number \tab Station \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab StationName \tab Station \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab StartTime \tab Station \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab Duration \tab Haul \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab Validity \tab Haul \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab Platform \tab Station \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab StartLatitude \tab Station \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab StartLongitude \tab Station \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab StopLatitude \tab Station \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab StopLongitude \tab Station \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab StatisticalRectangle \tab Station \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab MinTrawlDepth \tab Haul \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab MaxTrawlDepth \tab Haul \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab BottomDepth \tab Station \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab CruiseKey \tab Haul \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab StationKey \tab Haul \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab HaulKey \tab Haul \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab Number \tab Haul \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab Gear \tab Haul \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab Distance \tab Haul \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab Netopening \tab Haul \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab CodendMesh \tab Haul \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab SweepLength \tab Haul \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab GearExceptions \tab Haul \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab DoorType \tab Haul \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab WarpLength \tab Haul \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab WarpDiameter \tab Haul \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab WarpDensity \tab Haul \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab DoorSurface \tab Haul \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab DoorWeight \tab Haul \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab DoorSpread \tab Haul \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab WingSpread \tab Haul \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab Buoyancy \tab Haul \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab KiteArea \tab Haul \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab GroundRopeWeight \tab Haul \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab Rigging \tab Haul \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab Tickler \tab Haul \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab HydrographicStationID \tab Station \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab TowDirection \tab Station \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab SpeedGround \tab Station \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab SpeedWater \tab Station \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab WindDirection \tab Station \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab WindSpeed \tab Station \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab SwellDirection \tab Station \tab Haul \tab Station \cr
icesBiotic \tab Haul \tab SwellHeight \tab Station \tab Haul \tab Station \cr
icesBiotic \tab Catch \tab DataType \tab Sample \tab Sample \tab SpeciesCategory \cr
icesBiotic \tab Catch \tab CruiseKey \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
icesBiotic \tab Catch \tab StationKey \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
icesBiotic \tab Catch \tab HaulKey \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
icesBiotic \tab Catch \tab SpeciesCategoryKey \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
icesBiotic \tab Catch \tab SpeciesCode \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
icesBiotic \tab Catch \tab SpeciesValidity \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
icesBiotic \tab Catch \tab SpeciesCategory \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
icesBiotic \tab Catch \tab SpeciesCategoryNumber \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
icesBiotic \tab Catch \tab WeightUnit \tab Sample \tab Sample \tab SpeciesCategory \cr
icesBiotic \tab Catch \tab SpeciesCategoryWeight \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
icesBiotic \tab Catch \tab SpeciesSex \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
icesBiotic \tab Catch \tab CruiseKey \tab Sample \tab Sample \tab SpeciesCategory \cr
icesBiotic \tab Catch \tab StationKey \tab Sample \tab Sample \tab SpeciesCategory \cr
icesBiotic \tab Catch \tab HaulKey \tab Sample \tab Sample \tab SpeciesCategory \cr
icesBiotic \tab Catch \tab SpeciesCategoryKey \tab Sample \tab Sample \tab SpeciesCategory \cr
icesBiotic \tab Catch \tab SampleKey \tab Sample \tab Sample \tab SpeciesCategory \cr
icesBiotic \tab Catch \tab SpeciesCategoryWeight \tab Sample \tab Sample \tab SpeciesCategory \cr
icesBiotic \tab Catch \tab SpeciesCategoryNumber \tab Sample \tab Sample \tab SpeciesCategory \cr
icesBiotic \tab Catch \tab SubsampledNumber \tab Sample \tab Sample \tab SpeciesCategory \cr
icesBiotic \tab Catch \tab SubsamplingFactor \tab Sample \tab Sample \tab SpeciesCategory \cr
icesBiotic \tab Catch \tab SubsampleWeight \tab Sample \tab Sample \tab SpeciesCategory \cr
icesBiotic \tab Catch \tab LengthCode \tab Sample \tab Sample \tab SpeciesCategory \cr
icesBiotic \tab Catch \tab LengthClass \tab Sample \tab Sample \tab SpeciesCategory \cr
icesBiotic \tab Catch \tab LengthType \tab Sample \tab Sample \tab SpeciesCategory \cr
icesBiotic \tab Catch \tab NumberAtLength \tab Sample \tab Sample \tab SpeciesCategory \cr
icesBiotic \tab Catch \tab WeightAtLength \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.1 \tab fishstation \tab serialno \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab nation \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab platform.fishstation \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab station \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab fixedstation \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab startdate.fishstation \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab starttime \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab stopdate.fishstation \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab stoptime \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab stationtype \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab latitudestart \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab longitudestart \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab latitudeend \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab longitudeend \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab system \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab area \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab location \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab bottomdepthstart \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab bottomdepthstop \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab fishingdepthmax \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab fishingdepthmin \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab fishingdepthmean \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab gearno \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab gear \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab gearcount \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab direction \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab gearspeed \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab startlog \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab stoplog \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab distance \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab gearcondition \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab trawlquality \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab trawlopening \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab trawlopeningsd \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab doorspread \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab doorspreadsd \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab wirelength \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab soaktime \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab tripno \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab fishabundance \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab fishdistribution \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab landingsite \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab fishingground \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab countofvessels \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab dataquality \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab haulvalidity \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab flora \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab vegetationcover \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab visibility \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab flowcount \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab flowconst \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab fishstation \tab comment \tab Station \tab Haul \tab Station \cr
nmdbioticv1.1 \tab catchsample \tab species \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.1 \tab catchsample \tab samplenumber \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.1 \tab catchsample \tab noname \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.1 \tab catchsample \tab aphia \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.1 \tab catchsample \tab sampletype \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.1 \tab catchsample \tab group \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.1 \tab catchsample \tab conservation \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.1 \tab catchsample \tab producttype \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.1 \tab catchsample \tab weight \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.1 \tab catchsample \tab volume \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.1 \tab catchsample \tab count \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.1 \tab catchsample \tab abundancecategory \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.1 \tab catchsample \tab sampleproducttype \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.1 \tab catchsample \tab lengthmeasurement \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.1 \tab catchsample \tab lengthsampleweight \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.1 \tab catchsample \tab lengthsamplevolume \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.1 \tab catchsample \tab lengthsamplecount \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.1 \tab catchsample \tab specimensamplecount \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.1 \tab catchsample \tab raisingfactor \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.1 \tab catchsample \tab agingstructure \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.1 \tab catchsample \tab parasite \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.1 \tab catchsample \tab stomach \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.1 \tab catchsample \tab genetics \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.1 \tab catchsample \tab foreignobject \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.1 \tab catchsample \tab comment.catchsample \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.4 \tab fishstation \tab serialno \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab nation \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab platform.fishstation \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab station \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab fixedstation \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab startdate.fishstation \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab starttime \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab stopdate.fishstation \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab stoptime \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab stationtype \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab latitudestart \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab longitudestart \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab latitudeend \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab longitudeend \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab system \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab area \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab location \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab bottomdepthstart \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab bottomdepthstop \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab fishingdepthmax \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab fishingdepthmin \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab fishingdepthmean \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab gearno \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab gear \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab gearcount \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab direction \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab gearspeed \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab startlog \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab stoplog \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab distance \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab gearcondition \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab trawlquality \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab trawlopening \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab trawlopeningsd \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab doorspread \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab doorspreadsd \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab wirelength \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab soaktime \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab tripno \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab fishabundance \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab fishdistribution \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab landingsite \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab fishingground \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab countofvessels \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab dataquality \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab haulvalidity \tab Haul \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab flora \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab vegetationcover \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab visibility \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab flowcount \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab flowconst \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab waterlevel \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab winddirection \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab windspeed \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab clouds \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab sea \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab weather \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab fishstation \tab comment \tab Station \tab Haul \tab Station \cr
nmdbioticv1.4 \tab catchsample \tab species \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.4 \tab catchsample \tab samplenumber \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.4 \tab catchsample \tab noname \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.4 \tab catchsample \tab aphia \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.4 \tab catchsample \tab sampletype \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.4 \tab catchsample \tab group \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.4 \tab catchsample \tab conservation \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.4 \tab catchsample \tab producttype \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.4 \tab catchsample \tab weight \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.4 \tab catchsample \tab volume \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.4 \tab catchsample \tab count \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.4 \tab catchsample \tab abundancecategory \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.4 \tab catchsample \tab sampleproducttype \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.4 \tab catchsample \tab lengthmeasurement \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.4 \tab catchsample \tab lengthsampleweight \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.4 \tab catchsample \tab lengthsamplevolume \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.4 \tab catchsample \tab lengthsamplecount \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.4 \tab catchsample \tab specimensamplecount \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.4 \tab catchsample \tab raisingfactor \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.4 \tab catchsample \tab agingstructure \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.4 \tab catchsample \tab parasite \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.4 \tab catchsample \tab stomach \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.4 \tab catchsample \tab genetics \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.4 \tab catchsample \tab foreignobject \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
nmdbioticv1.4 \tab catchsample \tab comment.catchsample \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab fishstation \tab serialnumber \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab nation \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab catchplatform \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab station \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab fixedstation \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab stationstartdate \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab stationstarttime \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab stationstopdate \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab stationstoptime \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab stationtype \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab latitudestart \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab longitudestart \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab latitudeend \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab longitudeend \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab system \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab area \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab location \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab bottomdepthstart \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab bottomdepthstop \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab bottomdepthmean \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab fishingdepthstart \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab fishingdepthstop \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab fishingdepthcount \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab fishingdepthmax \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab fishingdepthmin \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab fishingdepthmean \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab fishingdepthtemperature \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab fishingbait \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab gearno \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab sweeplength \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab gear \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab gearcount \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab direction \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab gearflow \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab vesselspeed \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab logstart \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab logstop \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab distance \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab gearcondition \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab samplequality \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab verticaltrawlopening \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab verticaltrawlopeningsd \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab trawldoortype \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab trawldoorarea \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab trawldoorweight \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab trawldoorspread \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab trawldoorspreadsd \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab wingspread \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab wingspreadsd \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab wirelength \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab wirediameter \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab wiredensity \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab soaktime \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab tripno \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab fishabundance \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab fishdistribution \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab landingsite \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab fishingground \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab vesselcount \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab dataquality \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab haulvalidity \tab Haul \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab flora \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab vegetationcover \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab visibility \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab waterlevel \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab winddirection \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab windspeed \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab clouds \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab sea \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab weather \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab fishstation \tab stationcomment \tab Station \tab Haul \tab Station \cr
nmdbioticv3.1 \tab catchsample \tab commonname \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab catchcategory \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab catchpartnumber \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab aphia \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab scientificname \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab identification \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab foreignobject \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab sampletype \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab group \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab conservation \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab catchproducttype \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab raisingfactor \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab catchweight \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab catchvolume \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab catchcount \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab abundancecategory \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab sampleproducttype \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab lengthmeasurement \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab lengthsampleweight \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab lengthsamplevolume \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab lengthsamplecount \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab specimensamplecount \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab agesamplecount \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab agingstructure \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab parasite \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab stomach \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab intestine \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab tissuesample \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab samplerecipient \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab catchcomment \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3.1 \tab catchsample \tab catchsampleid \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab fishstation \tab serialnumber \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab nation \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab catchplatform \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab station \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab fixedstation \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab stationstartdate \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab stationstarttime \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab stationstopdate \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab stationstoptime \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab stationtype \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab latitudestart \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab longitudestart \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab latitudeend \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab longitudeend \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab system \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab area \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab location \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab bottomdepthstart \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab bottomdepthstop \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab bottomdepthmean \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab fishingdepthstart \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab fishingdepthstop \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab fishingdepthcount \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab fishingdepthmax \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab fishingdepthmin \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab fishingdepthmean \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab fishingdepthtemperature \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab gearno \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab sweeplength \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab gear \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab gearcount \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab direction \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab gearflow \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab vesselspeed \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab logstart \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab logstop \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab distance \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab gearcondition \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab samplequality \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab verticaltrawlopening \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab verticaltrawlopeningsd \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab trawldoortype \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab trawldoorarea \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab trawldoorweight \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab trawldoorspread \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab trawldoorspreadsd \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab wingspread \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab wingspreadsd \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab wirelength \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab wirediameter \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab wiredensity \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab soaktime \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab tripno \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab fishabundance \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab fishdistribution \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab landingsite \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab fishingground \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab vesselcount \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab dataquality \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab haulvalidity \tab Haul \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab flora \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab vegetationcover \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab visibility \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab waterlevel \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab winddirection \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab windspeed \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab clouds \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab sea \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab weather \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab fishstation \tab stationcomment \tab Station \tab Haul \tab Station \cr
nmdbioticv3 \tab catchsample \tab commonname \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab catchcategory \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab catchpartnumber \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab aphia \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab scientificname \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab identification \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab foreignobject \tab SpeciesCategory \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab sampletype \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab group \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab conservation \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab catchproducttype \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab raisingfactor \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab catchweight \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab catchvolume \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab catchcount \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab abundancecategory \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab sampleproducttype \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab lengthmeasurement \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab lengthsampleweight \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab lengthsamplevolume \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab lengthsamplecount \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab specimensamplecount \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab agesamplecount \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab agingstructure \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab parasite \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab stomach \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab intestine \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab tissuesample \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab samplerecipient \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab catchcomment \tab Sample \tab Sample \tab SpeciesCategory \cr
nmdbioticv3 \tab catchsample \tab catchsampleid \tab Sample \tab Sample \tab SpeciesCategory \cr
}
}
