% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{StoxBioticTranslation}
\alias{StoxBioticTranslation}
\title{Translation definition (from file) for \code{\link{StoxBioticData}}.}
\description{
Translation definition (from file) for \code{\link{StoxBioticData}}.
}
\details{
This StoX data type is produced by \code{\link{DefineStoxBioticTranslation}}, and contains the columns VariableName, Value and NewValue.
}
\seealso{
\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxData}}
}
