% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{StoxBioticVariableConversion}
\alias{StoxBioticVariableConversion}
\title{Conversion table for \code{\link{StoxBioticData}}.}
\description{
Conversion table for \code{\link{StoxBioticData}}.
}
\details{
This StoX data type is produced by \code{\link{DefineStoxBioticVariableConversion}}, and contains the columns VariableName, Value and NewValue.
}
\seealso{
\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxData}}
}
