% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoxFunctionAttributes.R
\name{stoxFunctionAttributes}
\alias{stoxFunctionAttributes}
\title{Function specification for inclusion in StoX projects}
\usage{
stoxFunctionAttributes
}
\description{
Function specification for inclusion in StoX projects
}
