% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineAndUpdateVariables.R
\name{TranslateAcoustic}
\alias{TranslateAcoustic}
\title{Translate AcousticData}
\usage{
TranslateAcoustic(AcousticData, Translation)
}
\arguments{
\item{AcousticData}{An input of \link{ModelData} object}

\item{Translation}{The process from which to get the \code{\link{Translation}} definition.}
}
\value{
A \code{\link{AcousticData}} object.
}
\description{
This function translates one or more columns of \code{\link{AcousticData}} to new values given by the input \code{Translation}.
}
