% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineAndUpdateVariables.R
\name{TranslateBiotic}
\alias{TranslateBiotic}
\title{Translate BioticData}
\usage{
TranslateBiotic(BioticData, Translation)
}
\arguments{
\item{BioticData}{An input of \link{ModelData} object}

\item{Translation}{The process from which to get the \code{\link{Translation}} definition.}
}
\value{
A \code{\link{BioticData}} object.
}
\description{
This function translates one or more columns of \code{\link{BioticData}} to new values given by the input \code{Translation}.
}
