% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwardCompatibility.R
\name{translateDatrasField}
\alias{translateDatrasField}
\title{Function to translate from old to new Datras field name:}
\usage{
translateDatrasField(x)
}
\arguments{
\item{x}{Character: A vector of strings.}
}
\description{
Function to translate from old to new Datras field name:
}
