% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineAndUpdateVariables.R
\name{TranslateICESAcoustic}
\alias{TranslateICESAcoustic}
\title{Translate ICESAcousticData}
\usage{
TranslateICESAcoustic(ICESAcousticData, Translation)
}
\arguments{
\item{ICESAcousticData}{An input of \link{ModelData} object}

\item{Translation}{The process from which to get the \code{\link{Translation}} definition.}
}
\value{
A \code{\link{ICESAcousticData}} object.
}
\description{
This function translates one or more columns of \code{\link{ICESAcousticData}} to new values given by the input \code{Translation}.
}
