% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineAndUpdateVariables.R
\name{TranslateICESDatras}
\alias{TranslateICESDatras}
\title{Translate ICESDatrasData}
\usage{
TranslateICESDatras(
  ICESDatrasData,
  TranslationDefinition = c("FunctionParameter", "FunctionInput"),
  VariableName = character(),
  Conditional = FALSE,
  ConditionalVariableNames = character(),
  TranslationTable = data.table::data.table(),
  Translation,
  PreserveClass = TRUE
)
}
\arguments{
\item{ICESDatrasData}{\code{\link{ICESDatrasData}}.}

\item{TranslationDefinition}{The method to use for defining the Translation, one of \code{FunctionParameter} to define the Translation on the fly in this function using the \code{TranslationTable}, or \code{FunctionInput} to import \code{\link{Translation}} process data from a previously run process using the function \code{\link{DefineTranslation}}. When reading the translation from a file \code{FunctionInput} must be used.}

\item{VariableName}{The name of the variable to translate.}

\item{Conditional}{Logical: If TRUE condition the translation on values of other variables.}

\item{ConditionalVariableNames}{The names of the variables to condition the translation on. Must be given if \code{Conditional == TRUE}.}

\item{TranslationTable}{A table holding the following columns: The first column holds the values to translate FROM, the second column holds the values to translate TO, and the remaining zero or more columns holds the values for the conditional variables specified in \code{ConditionalVariableNames}. I.e., if \code{VariableName} = "IndividualAge" and \code{ConditionalVariableNames} = "IndividualSex", and the \code{TranslationTable} has values 3, 4 and "F" in the first row, female fish at age 3 are translated to age 4. Use NA to translate missing values (shown as "-" in Preview in the StoX GUI, and usually as empty cell in excel). Values in the \code{TranslationTable} can be given either as single values or as expressions of functions of the variable specified by the column name. See details of \code{\link{DefineTranslation}}.}

\item{Translation}{\code{\link{Translation}}.}

\item{PreserveClass}{Logical: If TRUE (the default) do not convert the class of the data by the class of the translation table. E.g., with the default (\code{PreserveClass} = TRUE), if the variable to translate is integer (1, 2, etc) and the NewValue is character ("First", "Second", etc), the NewValue will be converted to integer before translation, resulting in NA if the character strings are not convertible to integer. In this example it could be the user's intention to convert the class of the variable to translate instead, which is possible using \code{PreserveClass} = FALSE.}
}
\value{
A \code{\link{ICESDatrasData}} object.
}
\description{
This function translates one or more columns of \code{\link{ICESDatrasData}} to new values given by the input \code{Translation}.
}
