% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineAndUpdateVariables.R
\name{TranslateStoxLanding}
\alias{TranslateStoxLanding}
\title{Translate StoxLandingData}
\usage{
TranslateStoxLanding(StoxLandingData, Translation)
}
\arguments{
\item{StoxLandingData}{An input of \link{ModelData} object}

\item{Translation}{The process from which to get the \code{\link{Translation}} definition.}
}
\value{
A \code{\link{StoxLandingData}} object.
}
\description{
This function translates one or more columns of \code{\link{StoxLandingData}} to new values given by the input \code{Translation}.
}
