% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Translation}
\alias{Translation}
\title{Translation definition (from file or from table).}
\description{
Translation definition (from file or from table).
}
\details{
This StoX data type is produced by \code{\link{DefineTranslation}}, and contains the columns VariableName, Value and NewValue.
}
\seealso{
\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxData}}
}
