% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeXmlFile.R
\name{unMergeAcoustic}
\alias{unMergeAcoustic}
\title{Un-merge MergeAcousticData}
\usage{
unMergeAcoustic(MergeAcousticData, useXsd, ...)
}
\arguments{
\item{MergeAcousticData}{The data to be split into \code{\link{AcousticData}}.}

\item{useXsd}{The xsd to use when splitting the flat tables into the different levels.}

\item{...}{Additionial values to inclcude in the output \code{\link{AcousticData}}. Applies to all tables in MergeAcousticData. The length of each of the elements given must match the number of rows of the table to add the values to.}
}
\description{
Function to split MergeAcousticData into \code{\link{AcousticData}}
}
