% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeXmlFile.R
\name{unMergeBiotic}
\alias{unMergeBiotic}
\title{Un-merge MergeBioticData}
\usage{
unMergeBiotic(MergeBioticData, useXsd, ...)
}
\arguments{
\item{MergeBioticData}{The data to be split into \code{\link{BioticData}}.}

\item{useXsd}{The xsd to use when splitting the flat tables into the different levels.}

\item{...}{Additionial values to inclcude in the output \code{\link{BioticData}}. Applies to all tables in MergeBioticData. The length of each of the elements given must match the number of rows of the table to add the values to.}
}
\description{
Function to split MergeBioticData into \code{\link{BioticData}}
}
