% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeXmlFile.R
\name{WriteBiotic}
\alias{WriteBiotic}
\title{Write Biotic}
\usage{
WriteBiotic(
  BioticData,
  FileNames = character(),
  namespaces = character(),
  encoding = "UTF-8",
  overwrite = FALSE
)
}
\arguments{
\item{BioticData}{\code{\link[RstoxData]{BioticData}} data to write.}

\item{FileNames}{paths to files that should be written to}

\item{namespaces}{XML namespaces to use for formatting.}

\item{encoding}{encoding to use for writing files}

\item{overwrite}{whether to overwrite any existing file(s)}
}
\description{
Write biotic data as XML file(s).
}
\details{
\code{\link[RstoxData]{BioticData}} may contain several data sets.
 In that case the parameters 'FileNames' and 'namespaces' must be provided for each
 in the order they appear in 'BioticData'.

Supports writing to namespaces:
  http://www.imr.no/formats/nmdbiotic/v1.1
  http://www.imr.no/formats/nmdbiotic/v1.2
  http://www.imr.no/formats/nmdbiotic/v1.3
  http://www.imr.no/formats/nmdbiotic/v3
  http://www.imr.no/formats/nmdbiotic/v3.1
}
\examples{
exampleFile <- system.file("testresources","biotic3.1_example.xml", package="RstoxData")
bioticData <- ReadBiotic(exampleFile)
newFile <- tempfile(fileext = ".xml")
WriteBiotic(bioticData, newFile, namespaces = "http://www.imr.no/formats/nmdbiotic/v3.1")
bioticDataReadBackIn <- ReadBiotic(newFile)
all.equal(
  bioticData[[1]][names(bioticData[[1]]) != "metadata"], 
  bioticDataReadBackIn[[1]][names(bioticDataReadBackIn[[1]]) != "metadata"]
)

}
