% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxExport.R
\name{writeICESAcoustic}
\alias{writeICESAcoustic}
\title{Write ICES acoustic CSV format file}
\usage{
writeICESAcoustic(Acoustic, save = TRUE)
}
\arguments{
\item{Acoustic}{a \code{AcousticData} object from an ICES acoustic XML format file.}

\item{save}{an output file in collated CSV format will be created if this parameter is set to TRUE.}
}
\value{
List of data.table objects in the ICES acoustic CSV format.
}
\description{
Given an \code{AcousticData} object, this function will write an ICES acoustic CSV file. Note that this function only supports
\code{AcousticData} object that is created from reading an ICES acoustic XML file.
}
