% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxExport.R
\name{WriteICESDatsusc}
\alias{WriteICESDatsusc}
\title{Write ICESDatras to CSV fille}
\usage{
WriteICESDatsusc(ICESDatsuscData)
}
\arguments{
\item{ICESDatsuscData}{\code{\link{ICESDatsuscData}}.}
}
\value{
An object of StoX data type \code{\link{WriteICESDatrasData}}.
}
\description{
Writes \code{\link{ICESDatrasData}} to a csv file for each input acoustic file used to create the \code{\link{ICESDatras}}
}
