% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xsdObjects}
\alias{xsdObjects}
\title{xsdObjects}
\format{
A list with 4 elements
\describe{
  \item{\code{landingerv2.xsd}}{List Landing Format v2}
  \item{\code{nmdbioticv1.xsd}}{List NMD Biotic Format v1}
  \item{\code{nmdbioticv1.1.xsd}}{List NMD Biotic Format v1.1}
  \item{\code{nmdbioticv1.2.xsd}}{List NMD Biotic Format v1.2}
  \item{\code{nmdbioticv1.3.xsd}}{List NMD Biotic Format v1.3}
  \item{\code{nmdbioticv1.4.xsd}}{List NMD Biotic Format v1.4}
  \item{\code{nmdbioticv3.xsd}}{List NMD Biotic Format v3}
  \item{\code{nmdbioticv3.1.xsd}}{List NMD Biotic Format v3.1}
  \item{\code{nmdechosounderv1.xsd}}{List NMD Echosounder Format v1}
}
}
\source{
\url{https://www.imr.no/formats}
}
\usage{
xsdObjects
}
\description{
Pre-processed XSD file objects
}
\keyword{datasets}
