% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBiotic.R
\name{AddToStoxBiotic}
\alias{AddToStoxBiotic}
\title{Add variables to StoxBioticData from BioticData}
\usage{
AddToStoxBiotic(StoxBioticData, BioticData, VariableNames = character())
}
\arguments{
\item{StoxBioticData}{\code{\link{StoxBioticData}}.}

\item{BioticData}{\code{\link{BioticData}}.}

\item{VariableNames}{A character vector with names of the variables to add from the \code{BioticData}.}
}
\value{
An object of StoX data type \code{\link{StoxBioticData}}.
}
\description{
Add variables to StoxBioticData from BioticData
}
