% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{applyFunctionArgumentHierarchy}
\alias{applyFunctionArgumentHierarchy}
\title{Function to return the names of the arguments to show for a function:}
\usage{
applyFunctionArgumentHierarchy(
  functionArgumentHierarchy,
  functionArguments,
  return.only.names = TRUE,
  ignore.condition = NULL
)
}
\arguments{
\item{functionArgumentHierarchy}{The function argument hierarchy defined in the stoxFunctionAttributes.}

\item{functionArguments}{A list of the arguments to the function (both function inputs and function parameters).}

\item{return.only.names}{Logical: If TRUE return only the names of the arguments to show.}

\item{ignore.condition}{Character: A vector of strings naming parameters to ignore in the hierarchy. Current use is "UseProcessData" to identify function inputs hidden by that parameter.}
}
\description{
Function to return the names of the arguments to show for a function:
}
