% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwardCompatibility.R
\docType{data}
\name{backwardCompatibility_RstoxData}
\alias{backwardCompatibility_RstoxData}
\title{Backward compabitibility actions:}
\format{
An object of class \code{list} of length 8.
}
\usage{
backwardCompatibility_RstoxData
}
\description{
Backward compabitibility actions:
}
\keyword{datasets}
