% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{BioticData}
\alias{BioticData}
\title{StoX data type BioticData}
\description{
Biotic data read from biotic xml files.
}
\details{

}
\seealso{
\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxData}}
}
