% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBiotic.R
\name{CompensateEffectiveTowDistanceForFishingDepthCount}
\alias{CompensateEffectiveTowDistanceForFishingDepthCount}
\title{Add variables to StoxBioticData from BioticData}
\usage{
CompensateEffectiveTowDistanceForFishingDepthCount(StoxBioticData)
}
\arguments{
\item{StoxBioticData}{\code{\link{StoxBioticData}}.}

\item{VariableNames}{A character vector with names of the variables to add from the \code{BioticData}.}
}
\value{
An object of StoX data type \code{\link{StoxBioticData}}.
}
\description{
Add variables to StoxBioticData from BioticData
}
