% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdirParsing.R
\name{convertToLandingData}
\alias{convertToLandingData}
\title{Convert landings}
\usage{
convertToLandingData(lssLandings)
}
\arguments{
\item{lssLandings}{landings in the format provided by \code{\link[RstoxData]{readLssFile}}}
}
\value{
\code{\link[RstoxData]{LandingData}} the converted landings
}
\description{
Convert landings to \code{\link[RstoxData]{LandingData}}.
}
\details{
When converting from lss:
 All columns of 'lssLandings' are converted, except 'salgsdato' and 'versjonstidspunkt'.
 All columns of \code{\link[RstoxData]{LandingData}} are filled, except 'DokumentFormulardato' and 'DokumentElektroniskDato'.
}
