% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{createOrderKey}
\alias{createOrderKey}
\title{Convert a vector to an integer vector where individual string elelents at interpreted as numeric if possible.}
\usage{
createOrderKey(x, split = "/")
}
\arguments{
\item{x}{A vector.}

\item{split}{A character to split strings by.}
}
\description{
Convert a vector to an integer vector where individual string elelents at interpreted as numeric if possible.
}
