% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineAndUpdateVariables.R
\name{DefineTranslation}
\alias{DefineTranslation}
\title{Define translation}
\usage{
DefineTranslation(
  processData,
  UseProcessData = FALSE,
  DefinitionMethod = c("ResourceFile", "Table"),
  TranslationTable = data.table::data.table(),
  VariableName = character(),
  ValueColumn = character(),
  NewValueColumn = character(),
  Conditional = FALSE,
  ConditionalVariableNames = character(),
  ConditionalValueColumns = character(),
  FileName = character()
)
}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{UseProcessData}{Logical: If TRUE use the existing function output in the process.}

\item{DefinitionMethod}{Character: A string naming the method to use, one of "Table" for defining the \code{Table}, and "ResourceFile" for reading the table from the file given by \code{FileName}.}

\item{TranslationTable}{A table of the variable to translate in the first column; the column \code{NewValue} giving the values to translate to in the second column; followed by zero or more conditional variables. Use NA in the Value column to translate missing values (shown as "-" in View output in the StoX GUI, and usually as empty cell in excel). Values in the \code{TranslationTable} can be given either as single values or as expressions or functions of the variable specified by the column name. See details.}

\item{VariableName}{An optional string naming the variable to translate, which can be given if only one variable should be translated. If more than one variable should be translated using the same \code{\link{Translation}}, a column named "VariableName" must be inclcuded in \code{TranslationTable} (possibly read from the \code{FileName}) AND \code{VariableName} must be an empty string (blank field in the GUI).}

\item{ValueColumn, NewValueColumn}{The name of the columns of \code{TranslationTable} representing the current values and the values to translate to, respectively.}

\item{Conditional}{Logical: If TRUE the columns \code{ConditionalVariableNames} and \code{ConditionalValue} are expected in the \code{TranslationTable}. These define a variable interacting with the \code{VariableName} and \code{Value}, so that \code{VariableName} is changed from \code{Value} to \code{NewValue} only when \code{ConditionalVariableNames} has the value given by \code{ConditionalValue}. Note that \code{ConditionalVariableNames} must exist in the same table as \code{VariableName}.}

\item{ConditionalVariableNames}{Similar to \code{VariableName}, but naming the conditional variables.}

\item{ConditionalValueColumns}{The name of the columns of \code{TranslationTable} representing the conditional values.}

\item{FileName}{The csv file holding a table with the three variables listed for \code{TranslationTable}, and possibly the two listed for \code{Conditional}.}
}
\value{
A \code{\link{Translation}} object.
}
\description{
This function defines the translation table used as input to \code{\link{TranslateStoxBiotic}} and similar functions to translate values of one or more columns to new values given by a table or read from a CSV file.
}
\details{
The columns of the \code{TranslationnTable} (excecpt the NewValue column) can be given in one of two ways: (1) A single value or a string to be evaluated and matched using the "\%in\%" operator, such as "HER" or "c(\"HER\", \"CLU\")"; or (2) a string expressing a function of the variable given by the column name, such as "function(IndividualTotalLength) IndividualTotalLength > 10". When the \code{TranslationnTable} is given in the StoX GUI the strings need not be escaped ((1) HER or c("HER", "CLU"); or (2) function(IndividualTotalLength) IndividualTotalLength > 10). 

E.g., to set all inidividuals with missing IndividualMaturity as "Adult" if longer than 10 cm, use "function(IndividualMaturity) is.na(IndividualMaturity)" in the first column named "IndividualMaturity", "Adult" in the "NewValue" column, and "function(IndividualTotalLength) IndividualTotalLength > 10" in the third (conditional) column named "IndividualTotalLength".
}
\seealso{
\code{\link{TranslateStoxBiotic}}, \code{\link{TranslateStoxAcoustic}}, \code{\link{TranslateBiotic}}, \code{\link{TranslateAcoustic}}, \code{\link{TranslateStoxLanding}}, \code{\link{TranslateLanding}}, \code{\link{TranslateICESBiotic}} and \code{\link{TranslateICESAcoustic}} for applying the translation.
}
