% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{FilterBiotic}
\alias{FilterBiotic}
\title{Filter (raw) Biotic data}
\usage{
FilterBiotic(BioticData, FilterExpression, FilterUpwards = FALSE)
}
\arguments{
\item{BioticData}{Input \code{\link{BioticData}} data.}

\item{FilterExpression}{Filter expression given as a list of strings. The name of the list and structures should be identical to the names of the input data list. To extract or exclude missing values (NAs) use the operator \code{\%in\%} or the special operator \code{\%notin\%}, which is defined in RstoxData.}

\item{FilterUpwards}{Whether the filter action will propagate in the upwards direction. Default to FALSE.}
}
\value{
An object of filtered data in the same format as the input data.
}
\description{
Filters \code{\link{BioticData}}.
}
