% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{FilterBiotic}
\alias{FilterBiotic}
\title{Filter (raw) Biotic data}
\usage{
FilterBiotic(
  BioticData,
  FilterExpression,
  PropagateDownwards = TRUE,
  PropagateUpwards = FALSE
)
}
\arguments{
\item{BioticData}{Input \code{\link{BioticData}} data.}

\item{FilterExpression}{Filter expression in list of strings. The name of the list and structures should be identical to the names of the input data list.}

\item{PropagateDownwards}{Whether the filter action will propagate in the downwards direction. Default to TRUE.}

\item{PropagateUpwards}{Whether the filter action will propagate in the upwards direction. Default to FALSE.}
}
\value{
An object of filtered data in the same format as the input data.
}
\description{
Filter (raw) Biotic data
}
