% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{filterData}
\alias{filterData}
\title{Run filter on any StoX related data source}
\usage{
filterData(
  inputData,
  filterExpression,
  propagateDownwards = TRUE,
  propagateUpwards = FALSE,
  ...
)
}
\arguments{
\item{inputData}{An input data. Can be a list of biotic data (StoX data type \code{\link{BioticData}}), list of acoustic data, StoxBiotic data, or StoxAcoustic data.}

\item{filterExpression}{Filter expression in list of strings. The name of the list and structures should be identical to the names of the input data list.}

\item{propagateDownwards}{Whether the filter action will propagate in the downwards direction. Default to TRUE.}

\item{propagateUpwards}{Whether the filter action will propagate in the upwards direction. Default to FALSE.}

\item{...}{Ignored}
}
\value{
An object of filtered data in the same format as the input data.
}
\description{
Applies filters to hierarchical data formats.
}
\details{
This is a convenient wrapper around \code{\link[RstoxData]{filterTables}}, which determines hierarchy based on available metadata.
For more explicit control of the hierarchical relations used, consider \code{\link[RstoxData]{filterTables}}.

Applies filters to a set of table that are hierarchically related, and provides options for
propagating filters up or down in the data hierarchy. Accepts either a list of data.tables or a
list of such lists.

Filters are specified as strings encoding logical expressions in standard R syntax, each associated to one of
the tables in the input.

After a filter is applied to a table, data that corresponds to the removed in other tables will be removed
as specified by the options propagateDownwards and propagateUpwards. These are defined in \code{\link[RstoxData]{filterTables}}:

Column names are assumed to be unique across all levels of the data hierarchy, except that each table is expected to
repeated the key-columns of all its ancestors.
}
\examples{
 filenames <- system.file("testresources",
       "biotic3.1_w_ageandprey.xml", package="RstoxData")
 inputData <- RstoxData:::ReadBiotic(filenames)
 
 #filters are nested list matching the structure of inputData
 filterExpression <- list()
 filterExpression$`biotic3.1_w_ageandprey.xml`$agedetermination <- c(
     'age < 10'
 )
 
 filteredData <- RstoxData::filterData(inputData, 
             filterExpression)
 nrow(filteredData$biotic3.1_w_ageandprey.xml$fishstation)
 filteredPropUp <- RstoxData::filterData(inputData, 
             filterExpression, 
             propagateUpwards = TRUE)
 nrow(filteredPropUp$biotic3.1_w_ageandprey.xml$fishstation)

}
