% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{filterData}
\alias{filterData}
\title{Convert BioticData to StoxBioticData}
\usage{
filterData(data, filterExpression, fropagateUpwards = "")
}
\arguments{
\item{BioticData}{A list of biotic data (StoX data type \code{\link{BioticData}}), one element for each input biotic file.}
}
\value{
An object of StoX data type \code{\link{StoxBioticData}}.
}
\description{
Convert BioticData to StoxBioticData
}
