% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{FilterICESDatsusc}
\alias{FilterICESDatsusc}
\title{Filter ICESDatras data}
\usage{
FilterICESDatsusc(ICESDatsuscData, FilterExpression = list())
}
\arguments{
\item{FilterExpression}{Filter expression in list of strings. The name of the list and structures should be identical to the names of the input data list.}

\item{ICESDatrasData}{Input \code{\link{ICESDatrasData}} data.}
}
\value{
An object of filtered data in the same format as the input data.
}
\description{
Filters \code{\link{ICESDatrasData}}.
}
