% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{getUnit}
\alias{getUnit}
\title{Get unit for value}
\usage{
getUnit(
  value,
  property = c("symbol", "name"),
  conversionTable = RstoxData::StoxUnits
)
}
\arguments{
\item{value}{value to read unit from}

\item{property}{the property of the unit that is to be returned (e.g. 'symbol' or 'name')}
}
\value{
the unit of value. <NA> if unit is not set
}
\description{
Get unit of a value in accordance with StoX convention.
}
\examples{
 dt <- data.table::data.table(weight=c(1000,1200))
 dt$weight <- setDtUnit(dt$weight, "g")
 print(getUnit(dt), "name")
}
