% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{getUnitOptions}
\alias{getUnitOptions}
\title{Get available units}
\usage{
getUnitOptions(
  quantity,
  property = c("shortname", "symbol", "name"),
  unitTable = RstoxData::StoxUnits,
  conversionRange = NULL
)

getUnitOptions(
  quantity,
  property = c("shortname", "symbol", "name"),
  unitTable = RstoxData::StoxUnits,
  conversionRange = NULL
)

getUnitOptions(
  quantity,
  property = c("shortname", "symbol", "name"),
  unitTable = RstoxData::StoxUnits,
  conversionRange = NULL
)
}
\arguments{
\item{quantity}{quantity to get units for, such as 'mass', 'length' etc.}

\item{property}{the property of the unit that is to be returned (e.g. 'symbol' or 'shortname')}

\item{unitTable}{formatted as \code{\link[RstoxData]{StoxUnits}}}

\item{conversionRange}{pair of numeric values providing the lower and upper bound of which values for unitTable$conversion should be provided}
}
\value{
a character vector with available unit symbols.

a character vector with availble unit symbols.

a character vector with available unit symbols.
}
\description{
Get the unit shortnames, symbols or names that are available for a given quantity.
 The available quantities and units are defined by the argument unitTable
 which defaults to RstoxData::StoxUnits
 
 This is useful for providing user options, and the options can be limited to a suitable range based on 'conversionRange'

Get the unit symbmols that are available for a given quantity.
 The available quntities and units are defined by the argument unitTable
 which defaults to RstoxData::StoxUnits

Get the unit shortnames, symbols or names that are available for a given quantity.
 The available quantities and units are defined by the argument unitTable
 which defaults to RstoxData::StoxUnits
}
\examples{
 print(getUnitOptions("mass"))
 print(getUnitOptions("mass", conversionRange=c(1e-3,1e3)))
 print(getUnitOptions("mass"))
 print(getUnitOptions("mass"))
 print(getUnitOptions("mass", conversionRange=c(1e-3,1e3)))
}
