% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxExport.R
\name{ICESAcoustic}
\alias{ICESAcoustic}
\title{Converts \code{AcousticData} to \code{ICESAcousticData}}
\usage{
ICESAcoustic(AcousticData)
}
\arguments{
\item{AcousticData}{A \code{AcousticData} object from an ICES acoustic XML format file.}
}
\value{
List of data.table objects in the ICES acoustic CSV format.
}
\description{
Note that this function only supports \code{AcousticData} object that is created from reading an ICES acoustic XML file.
}
