% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxExport.R
\name{ICESDatras}
\alias{ICESDatras}
\title{Write ICES DATRAS (NS-IBTS) format file}
\usage{
ICESDatras(BioticData)
}
\arguments{
\item{BioticData}{a \code{BioticData} object from an XML file with NMD biotic version 3 format.}
}
\value{
List of data.table objects in the ICES DATRAS CSV format.
}
\description{
Given an \code{BioticData} object, this function will write an ICES DATRAS (NS-IBTS) file. Note that this function only supports
\code{BioticData} object that is created from reading an NMD biotic version 3 XML file.
}
