% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{mapplyOnCores}
\alias{mapplyOnCores}
\title{Run a function on all elements of x on one or more cores}
\usage{
mapplyOnCores(
  FUN,
  NumberOfCores = integer(),
  ...,
  MoreArgs = NULL,
  SIMPLIFY = FALSE
)
}
\arguments{
\item{FUN}{The function to apply.}

\item{NumberOfCores}{The number of cores to use (defaulted to 1), truncated to the number of avaliable cores.}

\item{..., MoreArgs, SIMPLIFY}{See \code{\link[base]{mapply}}.}
}
\value{
A list of outputs from \code{FUN}.
}
\description{
Run a function on all elements of x on one or more cores
}
