% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{mergeByIntersect}
\alias{mergeByIntersect}
\title{Merge two data tables by the intersect of the names}
\usage{
mergeByIntersect(x, y, ...)
}
\arguments{
\item{x, y}{Data tables of class \code{\link[data.table]{data.table}}).}
}
\value{
A merged data table.
}
\description{
Merge two data tables by the intersect of the names
}
