% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{mergeByStoxKeys}
\alias{mergeByStoxKeys}
\title{Merge two data tables by StoX keys}
\usage{
mergeByStoxKeys(x, y, StoxDataType, toMergeFromY = NULL, replace = FALSE, ...)
}
\arguments{
\item{x, y}{Data tables of class \code{\link[data.table]{data.table}}.}

\item{StoxDataType}{Input data type. Text string of \code{StoxBiotic}
or \code{StoxAcoustic}.}

\item{toMergeFromY}{Specify key columns from \code{y}. \code{NULL} means
all similarly named columns from \code{x} and \code{y} will be
merged. Default to \code{NULL}.}

\item{replace}{Whether to replace the variables in the target.
Default to \code{FALSE}.}

\item{...}{Extra parameters that will be passed into \code{\link[data.table]{merge}}.}
}
\value{
A merged data table.
}
\description{
Merge two data tables by StoX keys
}
