% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{mergeDataTables}
\alias{mergeDataTables}
\title{Merge data tables}
\usage{
mergeDataTables(data, tableNames = NULL)
}
\arguments{
\item{data}{A list of data tables.}

\item{tableNames}{A character vector holding the names of the tables to merge.}
}
\value{
A merged data table.
}
\description{
Merge data tables
}
