% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAcoustic.r
\name{MergeStoxAcoustic}
\alias{MergeStoxAcoustic}
\title{Merge StoxAcousticData}
\usage{
MergeStoxAcoustic(StoxAcousticData, TargetTable = "NASC")
}
\arguments{
\item{StoxBioticData}{A list of StoX acoustic data (StoX data type \code{\link{StoxAcousticData}}).}
}
\value{
An object of StoX data type \code{\link{MergedStoxAcousticData}}.
}
\description{
Merge StoxAcousticData
}
