% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ModelData}
\alias{ModelData}
\title{StoX data types of the RstoxData package}
\arguments{
\item{BioticData}{\code{\link{BioticData}}.}

\item{StoxBioticData}{\code{\link{StoxBioticData}}.}

\item{AcousticData}{\code{\link{AcousticData}}.}

\item{StoxAcousticData}{\code{\link{StoxAcousticData}}.}
}
\description{
StoX data types are the data types used to transfer data and information between processes in a StoX estimation model.
}
\details{
This RstoxData package produces the following StoX data types:
\itemize{
\item{\code{\link{BioticData}}}
\item{\code{\link{StoxBioticData}}}
\item{\code{\link{MergeStoxBioticData}}}
\item{\code{\link{AcousticData}}}
\item{\code{\link{StoxAcousticData}}}
\item{\code{\link{MergeStoxAcousticData}}}
\item{\code{\link{LandingData}}}
\item{\code{\link{StoxLandingData}}}
}
}
\seealso{
\href{https://github.com/StoXProject/RstoxBase}{RstoxBase} and \href{https://github.com/StoXProject/RstoxFDA}{RstoxFDA} for a list of all StoX data types produced by the other official StoX function packages.
}
