% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ProcessData}
\alias{ProcessData}
\title{Process data used by RstoxData}
\arguments{
\item{Translation}{\code{\link{Translation}}.}
}
\description{
The process data of a StoX model are data that are saved to the project description (project.json file), which in the case of RstoxData is only the \code{\link{Translation}} data type.
}
