% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.R
\name{ReadAcoustic}
\alias{ReadAcoustic}
\title{Read acoustic XML files}
\usage{
ReadAcoustic(FileNames, NumberOfCores = 1L)
}
\arguments{
\item{FileNames}{The paths of the acoustic files.}

\item{NumberOfCores}{The number of cores to use (defaulted to 1), truncated to the number of avaliable cores.}
}
\value{
An object of StoX data type AcousticData: A list of a list of data.tables of the different levels of the input acoustic files.
}
\description{
This function reads multiple acoustic file to a list with a list of tables for each file.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link[RstoxData]{readXmlFile}}.
}
