% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineAndUpdateVariables.R
\name{RedefineStoxBiotic}
\alias{RedefineStoxBiotic}
\title{Redefine StoxBioticData variables by data from BioticData}
\usage{
RedefineStoxBiotic(
  StoxBioticData,
  BioticData,
  Redefinition = data.table::data.table(),
  SplitTableAllocation = c("Default", "Lowest", "Highest")
)
}
\arguments{
\item{StoxBioticData}{An input of \link{ModelData} object}

\item{BioticData}{An input of \link{ModelData} object}

\item{Redefinition}{A table of the columns "VariableName", representing the variable to redefine; and "ReplaceBy", representing the variable from BioticData to replace by.}

\item{SplitTableAllocation}{A string indicating how to split tables of the BioticData into tables of the StoxBiotic format, one of "Default" for the default mapping, "Lowest" for mapping variables to the lowest table (e.g. mapping variables from the fishstation level of \code{BioticData} from NMDBiotic files to the Haul level of \code{StoxBioticData}) and "Highest" for mapping variables to the highest table (e.g. mapping variables from the fishstation level of \code{BioticData} from NMDBiotic files to the Station level of \code{StoxBioticData}). See \code{\link{StoxBioticMapping}} for a description of the mapping for the different biotic input file formats.}
}
\value{
A \code{\link{StoxBioticData}} object.
}
\description{
This function redefines one or more columns of \code{\link{StoxBioticData}} by columns of \code{\link{BioticData}}.
}
