% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{RstoxData}
\alias{RstoxData}
\alias{RstoxData-package}
\title{R libraries to parse and write fisheries biotic and acoustic data format}
\description{
Tools to fetch and manipulate various data formats for fisheries (mainly geared towards biotic and acoustic data).
}
\details{
The RstoxData package contains functions for reading, filtering and writing biotic, acoustic and landing data as XML files. Filtering can be done by R syntax such as longitude > 10, or by pre defined functions such as inside().
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StoXProject/RstoxData}
  \item Report bugs at \url{https://github.com/StoXProject/RstoxData/issues}
}

}
\author{
\strong{Maintainer}: Ibrahim Umar \email{ibrahim.umar@hi.no}

Authors:
\itemize{
  \item Sindre Vatnehol
  \item Arne Johannes Holmin
  \item Edvin Fuglebakk
}

}
