% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{setColumnClasses}
\alias{setColumnClasses}
\title{Convert classes of a data.table:}
\usage{
setColumnClasses(DT, classes = NULL)
}
\arguments{
\item{DT}{A data.table}

\item{classes}{A list of classes named by the columns for which the class should be set. Columns in \code{DT} with name found in the names of \code{classes} are modified if the existing class differs from the desired class.}
}
\description{
Convert classes of a data.table:
}
