% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{setorderv_numeric}
\alias{setorderv_numeric}
\title{Order a data.table (by reference) by interpreting characters as numeric if possible}
\usage{
setorderv_numeric(dataOne, by = NULL, key = NULL, ...)
}
\arguments{
\item{dataOne}{A data.table.}

\item{by}{Order by the given columns.}

\item{key}{If given and \code{by} is empty, orderr by the columns with names ending with \code{key}.}

\item{...}{Passed on to \code{\link[data.table]{setorderv}}}
}
\description{
Order a data.table (by reference) by interpreting characters as numeric if possible
}
