% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{setorderv_numeric}
\alias{setorderv_numeric}
\title{Order a data.table (by reference) by interpreting characters as numeric if possible}
\usage{
setorderv_numeric(dataOne, by = NULL, key = NULL, ...)
}
\arguments{
\item{data}{A data.table.}
}
\description{
Order a data.table (by reference) by interpreting characters as numeric if possible
}
