% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{setRstoxPrecisionLevel}
\alias{setRstoxPrecisionLevel}
\title{Round off to number of digits}
\usage{
setRstoxPrecisionLevel(x)
}
\arguments{
\item{x}{A list of \code{data.table}s or a single \code{data.table} object.}
}
\value{
A transformed object.
}
\description{
Round off to number of digits
}
