% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{setUnit}
\alias{setUnit}
\title{Set units of value}
\usage{
setUnit(value, desired, conversionTable = RstoxData::StoxUnits)
}
\arguments{
\item{value}{value to set unit for.}

\item{desired}{the desired unit for the column}

\item{conversionTable}{formatted as \code{\link[RstoxData]{StoxUnits}}}
}
\value{
converted value with the attribute 'Unit' set / altered.
}
\description{
Set unit of a value in accordance with StoX convention (adds the attribute 'stoxUnit').
 Converts unit if already set.
}
\details{
The argument 'conversionTable' defines valid units and their symbols.
 This defaults to \code{\link[RstoxData]{StoxUnits}}.
}
\examples{
 dt <- data.table::data.table(weight=c(1000,1200))
 dt$weight <- setUnit(dt$weight, "g")
 print(dt$weight)
 dt$weight <- setUnit(dt$weight, "kg")
 print(dt$weight)
}
