% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAcoustic.R
\name{StoxAcoustic}
\alias{StoxAcoustic}
\title{Convert AcousticData to StoxAcousticData}
\usage{
StoxAcoustic(AcousticData, NumberOfCores = 1L)
}
\arguments{
\item{AcousticData}{\code{\link{AcousticData}}.}

\item{NumberOfCores}{The number of cores to use (defaulted to 1), truncated to the number of avaliable cores.}
}
\value{
An object of StoX data type \code{\link{StoxAcousticData}}.
}
\description{
Convert AcousticData to StoxAcousticData
}
