% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAcoustic.R
\name{StoxAcoustic}
\alias{StoxAcoustic}
\title{Convert AcousticData to StoxAcousticData}
\usage{
StoxAcoustic(AcousticData, NumberOfCores = integer())
}
\arguments{
\item{AcousticData}{A list of acoustic data (StoX data type \code{\link{AcousticData}}), one element for each input acoustic file.}

\item{NumberOfCores}{Overrides multi-core auto detection (default).}
}
\value{
An object of StoX data type \code{\link{StoxAcousticData}}.
}
\description{
Convert AcousticData to StoxAcousticData
}
