% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAcoustic.R
\name{StoxAcoustic}
\alias{StoxAcoustic}
\title{Convert AcousticData to StoxAcousticData}
\usage{
StoxAcoustic(AcousticData)
}
\arguments{
\item{AcousticData}{\code{\link{AcousticData}}.}
}
\value{
An object of StoX data type \code{\link{StoxAcousticData}}.
}
\description{
Convert AcousticData to StoxAcousticData
}
