% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAcoustic.r
\name{StoxAcoustic}
\alias{StoxAcoustic}
\title{Convert output from xml reader to StoxAcoustic}
\usage{
StoxAcoustic(data_list = NULL, converter = NULL)
}
\arguments{
\item{data_list}{table. The output from xmlreader.}

\item{converter}{file. Instruction file to convert values}
}
\value{
List of data.table objects containing the "flattened" XML data.
}
\description{
Convert output from xml reader to StoxAcoustic
}
