% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAcoustic.r
\name{StoxAcoustic}
\alias{StoxAcoustic}
\title{Convert output from xml reader to StoxAcoustic data}
\usage{
StoxAcoustic(AcousticData = NULL)
}
\arguments{
\item{AcousticData}{The output from xmlreader.}
}
\value{
List of data.table objects containing the StoxAcoustic data.
}
\description{
Convert output from xml reader to StoxAcoustic data
}
