% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBiotic.R
\name{StoxBiotic}
\alias{StoxBiotic}
\title{Convert BioticData to StoxBioticData}
\usage{
StoxBiotic(BioticData, NumberOfCores = 1L)
}
\arguments{
\item{BioticData}{\code{\link{BioticData}}.}

\item{NumberOfCores}{The number of cores to use (defaulted to 1), truncated to the number of avaliable cores.}
}
\value{
An object of StoX data type \code{\link{StoxBioticData}}.
}
\description{
Convert BioticData to StoxBioticData
}
