% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBiotic.R
\name{StoxBiotic}
\alias{StoxBiotic}
\title{Convert BioticData to StoxBioticData}
\usage{
StoxBiotic(BioticData)
}
\arguments{
\item{BioticData}{\code{\link{BioticData}}.}
}
\value{
An object of StoX data type \code{\link{StoxBioticData}}.
}
\description{
Convert BioticData to StoxBioticData
}
\seealso{
The definition of the \code{\link[=StoxBioticFormat]{StoxBiotic format}} and \code{\link{generalSamplingHierarhcy}}.
}
