% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBiotic.R
\name{StoxBiotic}
\alias{StoxBiotic}
\title{Convert BioticData to StoxBioticData}
\usage{
StoxBiotic(BioticData, Cores = integer())
}
\arguments{
\item{BioticData}{A list of biotic data (StoX data type \code{\link{BioticData}}), one element for each input biotic file.}

\item{Cores}{Overrides multi-core auto detection. Default to NULL.}
}
\value{
An object of StoX data type \code{\link{StoxBioticData}}.
}
\description{
Convert BioticData to StoxBioticData
}
