% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBiotic.R
\name{StoxBiotic}
\alias{StoxBiotic}
\title{Convert BioticData to StoxBioticData}
\usage{
StoxBiotic(BioticData)
}
\arguments{
\item{BioticData}{A list of biotic data (StoX data type \code{\link[=StoxBiotic]{BioticData}}), one element for each input biotic file.}
}
\value{
An object of StoX data type \code{\link[=StoxBiotic]{StoxBioticData}}.
}
\description{
Convert BioticData to StoxBioticData
}
