% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoxFunctionAttributes.R
\docType{data}
\name{stoxFunctionAttributes}
\alias{stoxFunctionAttributes}
\title{Function specification for inclusion in StoX projects}
\format{
An object of class \code{list} of length 19.
}
\usage{
stoxFunctionAttributes
}
\description{
Function specification for inclusion in StoX projects
}
\keyword{datasets}
