% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoxLanding.R
\name{StoxLanding}
\alias{StoxLanding}
\title{Convert landing data}
\usage{
StoxLanding(LandingData)
}
\arguments{
\item{LandingData}{Sales-notes data. See \code{\link[RstoxData]{LandingData}}}
}
\value{
\code{\link[RstoxData]{StoxLandingData}}, aggregated sales-notes data.
}
\description{
StoX function
 Convert landing data to the aggregated format \code{\link[RstoxData]{StoxLandingData}}
}
