% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxExport.R
\name{WriteICESAcoustic}
\alias{WriteICESAcoustic}
\title{Write ICES acoustic CSV format file}
\usage{
WriteICESAcoustic(AcousticData, Combine = FALSE)
}
\arguments{
\item{AcousticData}{A \code{AcousticData} object from an ICES acoustic XML format file.}

\item{Combine}{Logical: If TRUE stack the output tables for each acosutic file.}
}
\value{
List of data.table objects in the ICES acoustic CSV format.
}
\description{
Given an \code{AcousticData} object, this function will write an ICES acoustic CSV file. Note that this function only supports
\code{AcousticData} object that is created from reading an ICES acoustic XML file.
}
