% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxExport.R
\name{WriteICESAcoustic}
\alias{WriteICESAcoustic}
\title{Writes \code{\link{ICESAcousticData}} to a csv file for each input acoustic file used to create the \code{\link{ICESAcousticData}}}
\usage{
WriteICESAcoustic(ICESAcousticData)
}
\arguments{
\item{ICESAcousticData}{A \code{\link{ICESAcousticData}} object obtained from an ICES acoustic XML format file.}
}
\value{
List of string matrices in the ICES acoustic CSV format.
}
\description{
Writes \code{\link{ICESAcousticData}} to a csv file for each input acoustic file used to create the \code{\link{ICESAcousticData}}
}
