% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxExport.R
\name{WriteICESAcoustic}
\alias{WriteICESAcoustic}
\title{Write ICESAcoustic to CSV file}
\usage{
WriteICESAcoustic(ICESAcousticData, AggregateSa = FALSE)
}
\arguments{
\item{ICESAcousticData}{A \code{\link{ICESAcousticData}} object obtained from an ICES acoustic XML format file.}

\item{AggregateSa}{Logical: If TRUE aggregate the acoustic value for each LogDistance, SampleChannelDepthUpper and DataSaCategory, which should be done if multiple DataSaCategory have been translated to the same DataSaCategory.}
}
\value{
An object of StoX data type \code{\link{WriteICESAcousticData}}.
}
\description{
Writes \code{\link{ICESAcousticData}} to a csv file for each input acoustic file used to create the \code{\link{ICESAcousticData}}
}
