% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxExport.R
\name{WriteICESBiotic}
\alias{WriteICESBiotic}
\title{Writes \code{\link{ICESBioticData}} to a csv file for each input acoustic file used to create the \code{\link{ICESBioticData}}}
\usage{
WriteICESBiotic(ICESBioticData)
}
\arguments{
\item{ICESBioticData}{A \code{\link{ICESBioticData}} object obtained from an ICES acoustic XML format file.}
}
\value{
List of string matrices in the ICES acoustic CSV format.
}
\description{
Writes \code{\link{ICESBioticData}} to a csv file for each input acoustic file used to create the \code{\link{ICESBioticData}}
}
