% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{WriteICESBioticData}
\alias{WriteICESBioticData}
\title{Rbind \code{\link{ICESBioticData}} to a string matrix.}
\description{
The output of this function is suited for submission to \url{https://acoustic.ices.dk/}.
}
\details{
The ICESBiotic CSV format is one string matrix containing all tables of \code{\link{ICESBioticData}}, where column names are inclcuded as header rows.
}
\seealso{
\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxData}}
}
