% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxExport.R
\name{WriteICESDatras}
\alias{WriteICESDatras}
\title{Writes \code{\link{ICESDatrasData}} to a csv file for each input acoustic file used to create the \code{\link{ICESDatras}}}
\usage{
WriteICESDatras(ICESDatrasData)
}
\arguments{
\item{ICESDatrasData}{A \code{\link{ICESDatrasData}} object returned from \code{\link{ICESDatras}}.}
}
\value{
List of string matrices in the ICES Datras CSV format.
}
\description{
Writes \code{\link{ICESDatrasData}} to a csv file for each input acoustic file used to create the \code{\link{ICESDatras}}
}
