% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{WriteICESDatsuscData}
\alias{WriteICESDatsuscData}
\title{Rbind \code{\link{ICESDatsuscData}} to a string matrix.}
\description{
The output of this function is suited for submission to \url{https://www.ices.dk/data/data-portals/Pages/Stomach-content.aspx}.
}
\details{
The ICESDatras CSV format is one string matrix containing all tables of \code{\link{ICESDatsuscData}}, where column names are inclcuded as header rows.
}
\seealso{
\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxData}}
}
